==> 01_exclusion.Rmd <==
---
title: "Exclusion"
output: 
  html_document:
    toc: true
    toc_depth: 4
    toc_float: true
    code_folding: 'show'
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(warning = FALSE)
```

## Library
```{r}
library(formr)
library(dplyr)
library(tidyr)
```

## Load data
```{r}
data_wrangled = read.csv(file = "data/data_wrangled.csv")[,-1]
```

## Total number of responses 
```{r}
nrow(data_wrangled) # total number of clicks on the survey
nrow(data_wrangled %>% filter(age >= 18)) # total number of responses for participants 18 or older
```
94,738 clicks on the survey in total. 68,085 responses from participants 18 years or older. 


## Add id variable to wrangled data
```{r}
data_wrangled$id = 1:94738
```


## Exclusion criteria
### For main analyses
#### 1. Excluding participants not identifying as women
```{r}
data_wrangled_main = data_wrangled %>%
  filter(sex == "Woman")

nrow(data_wrangled_main) # 64879 women 
nrow(data_wrangled) - nrow(data_wrangled_main) # number of excluded participants

table(data_wrangled_main$sex) # check gender
table(data_wrangled_main$sexual_orientation) # check sexual orientations
```
N = 29,859 participants were excluded, leaving N = 64,879 participants.


#### 2. Excluding participants not identifying as asexual or heterosexual
```{r}
data_wrangled_main = data_wrangled_main %>% 
  filter(sexual_orientation == "Straight/Heterosexual" | sexual_orientation == "Asexual")

nrow(data_wrangled_main) # 53390 women 
64879 - nrow(data_wrangled_main) # number of excluded participants
table(data_wrangled_main$sexual_orientation) # 457 asexuals and 52933 heterosexuals
``` 
N = 11,489 participants were excluded, leaving N = 53,390 participants. Of those, n = 457 identify as asesxual and n = 52,933 identify as heterosexual. 


#### 3. Excluding participants choosing not to answer the honesty question or indicating not having filled out the survey truthfully
```{r}
data_wrangled_main = data_wrangled_main %>% 
 mutate(answer_accuracy = replace_na(answer_accuracy, "No value"))
data_wrangled_main = data_wrangled_main %>%
  filter(answer_accuracy == "I took the survey seriously; please use my information in the study." | answer_accuracy == "No value")

nrow(data_wrangled_main) # 51775 women 
53390 - nrow(data_wrangled_main) # number of excluded participants
table(data_wrangled_main$sexual_orientation) # 447 asexuals and 51328 heterosexuals

# check NAs
table(is.na(data_wrangled_main$answer_accuracy)) 
table(data_wrangled_main$answer_accuracy)
```
N = 1615 participants were excluded, leaving N = 51,775 participants. Of those, n = 447 identify as asexual and n = 51,328 identify as heterosexual.


### For robustness check
#### 1. Excluding participants not identifying as women, man or genderqueer/nonbinary
```{r}
data_wrangled_robustness = data_wrangled %>%
  filter(sex == "Woman" | sex == "Man" | sex == "Genderqueer/Nonbinary")

nrow(data_wrangled_robustness) # 67774 participants 
nrow(data_wrangled) - nrow(data_wrangled_robustness) # number of excluded participants

table(data_wrangled_robustness$sex) # check gender
table(data_wrangled_robustness$sexual_orientation)# check sexual orientations
```
N = 26,964 participants were excluded, leaving N = 67,774 participants. Of those, n = 64,879 identify as women, n = 2055 as men, and n = 840 as genderqueer/nonbinary.


#### 2. Excluding participants not identifying as asexual or heterosexual
```{r}
data_wrangled_robustness = data_wrangled_robustness %>% 
  filter(sexual_orientation == "Straight/Heterosexual" | sexual_orientation == "Asexual")

nrow(data_wrangled_robustness) # 55076 participants
67774 - nrow(data_wrangled_robustness) # number of excluded participants

table(data_wrangled_robustness$sex) # check gender
table(data_wrangled_robustness$sexual_orientation) # check sexual orientation
```
N = 12,698 participants were excluded, leaving N = 55,076 participants. Of those, n = 53,390 identify as women, n = 1581 as men, and n = 105 as genderqueer/nonbinary.
N = 539 identify as asexual and n = 54,537 as heterosexual.


#### 3. Excluding participants choosing not to answer the honesty question or indicating not having filled out the survey truthfully
```{r}
data_wrangled_robustness = data_wrangled_robustness %>% 
 mutate(answer_accuracy = replace_na(answer_accuracy, "No value"))
data_wrangled_robustness = data_wrangled_robustness %>%
  filter(answer_accuracy == "I took the survey seriously; please use my information in the study." | answer_accuracy == "No value")

nrow(data_wrangled_robustness) # 53412 partiicpants
55076 - nrow(data_wrangled_robustness) # total number of excluded participants

table(data_wrangled_robustness$sex) # check gender
table(data_wrangled_robustness$sexual_orientation) # check sexual orientation
```
N = 1664 participants were excluded, leaving N = 53,412 participants. Of those, n = 51,775 identify as women, n = 1535 as men, and n = 102 as genderqueer/nonbinary.
N = 527 identify as asexual and n = 52,885 as heterosexual. 


## Total number of excluded participants
```{r}
# for main analyses
nrow(data_wrangled) - nrow(data_wrangled_main)

# for robustness check
nrow(data_wrangled) - nrow(data_wrangled_robustness)
```
For the main analyses, a total of 42,963 participants were excluded, leaving 51,775 participants.
For the robustness analyses, a total of 41,326 participants were excluded, leaving 53,412 participants. 

## Included participants
```{r}
# for main analyses
# by sexual orientation
table(data_wrangled_main$sexual_orientation)

# for robustness check
# by sexual orientation
table(data_wrangled_robustness$sexual_orientation)

# by gender
table(data_wrangled_robustness$sex)

# by gender and sexual orientation
xtabs(~sex+sexual_orientation, data_wrangled_robustness) 
```


## Save data
### For main analyses
```{r}
write.csv(data_wrangled_main,
          file = "data/data_main.csv")
```

### For robustness check
```{r}
write.csv(data_wrangled_robustness,
          file = "data/data_robustness.csv")
```
