==> 03_codebook_robustness.Rmd <==
---
title: "Codebook Robustness"
output:
  html_document:
    toc: true
    toc_depth: 4
    toc_float: true
    code_folding: 'hide'
    self_contained: true
  pdf_document:
    toc: yes
    toc_depth: 4
    latex_engine: xelatex
editor_options: 
  chunk_output_type: console
---


## Setting options
```{r setup}
knitr::opts_chunk$set(
  warning = TRUE, # show warnings during codebook generation
  message = TRUE, # show messages during codebook generation
  error = TRUE, # do not interrupt codebook generation in case of errors,
                # usually better for debugging
  echo = TRUE  # show R code
)
ggplot2::theme_set(ggplot2::theme_bw())

```

## Load library
```{r}
library(codebook)
library(dplyr)
```

## Prepare data for codebook
```{r prepare_codebook}
codebook_data_robustness = read.csv(file = "data/data_robustness.csv")[,-1]
```


## Check variable types and convert if necessary
```{r}
# sex should be a factor 
typeof(codebook_data_robustness$sex)
codebook_data_robustness$sex = as.factor(codebook_data_robustness$sex)

# sexual orientation should be a factor
typeof(codebook_data_robustness$sexual_orientation)
codebook_data_robustness$sexual_orientation = as.factor(codebook_data_robustness$sexual_orientation)

# country should be a factor
typeof(codebook_data_robustness$country)
codebook_data_robustness$country = as.factor(codebook_data_robustness$country)

# language should be a factor
typeof(codebook_data_robustness$language)
codebook_data_robustness$language = as.factor(codebook_data_robustness$language)

# transgender should be a factor
typeof(codebook_data_robustness$transgender)
codebook_data_robustness$transgender = as.factor(codebook_data_robustness$transgender)

# relationship status should be a factor
typeof(codebook_data_robustness$relationship_status)
codebook_data_robustness$relationship_status = as.factor(codebook_data_robustness$relationship_status)

# marital status should be a factor
typeof(codebook_data_robustness$marital_status)
codebook_data_robustness$marital_status = as.factor(codebook_data_robustness$marital_status)
```



## Create codebook for robustness data set
```{r}
codebook_robustness = codebook_data_robustness %>% select(sexual_orientation, country, language, age, sex, 
                                                                   transgender, relationship_length, relationship_status,
                                                                   marital_status, political_spectrum,
                                                                   interest_hookups,
                                                                   interest_nonsexrel,
                                                                   interest_monrel,
                                                                   interest_monrel,
                                                                   interest_altrel,
                                                                   interest_single,
                                                                   interest_parent,
                                                                   pref_imp_ks,
                                                                   pref_imp_att,
                                                                   pref_imp_fs,
                                                                   pref_imp_ca,
                                                                   pref_imp_ei,
                                                                   pref_imp_sexually_experienced,
                                                                   self_ks,
                                                                   self_att,
                                                                   self_fs,
                                                                   self_ca,
                                                                   self_ei,
                                                                   self_sexually_experienced)
                                        

codebook(codebook_robustness)
```
