==> 04_descriptives_before_matching.Rmd <==
---
title: "Descriptives Before Matching"
output: 
  html_document:
    toc: true
    toc_depth: 4
    toc_float: true
    code_folding: 'show'
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(warning = FALSE)
```

## Library
```{r}
library(dplyr)
library(ggplot2)
library(psych)
library(apaTables)
```

## Load data
```{r}
data_main = read.csv(file = "data/data_main.csv")[,-1]
```

## Descriptives for whole sample
### Demographic data
#### Sexual orientation
```{r}
table(data_main$sexual_orientation) # sexual orientation in absolute numbers
round(table(data_main$sexual_orientation)/sum(table(data_main$sexual_orientation)),3)*100 # sexual orientation in percent
table(is.na(data_main$sexual_orientation)) # no NAs 
```

#### Country
```{r}
table(data_main$country) # country in absolute numbers
round(table(data_main$country)/sum(table(data_main$country)),4)*100 # country in percent
table(is.na(data_main$country)) # no NAs 
```

#### Language
Chinese, Danish, English, French, German, Japanese, Portuguese, Russian, Spanish, Italian
```{r Language}
table(data_main$language) # language in absolute numbers
round(table(data_main$country)/sum(table(data_main$country)),4)*100 # country in percent
table(is.na(data_main$language)) #no NAs
```

#### Age
```{r}
table(data_main$age)
range(data_main$age, na.rm = T) # range
table(is.na(data_main$age)) # NAs

# histogram 
hist(data_main$age)

# mean, sd
round(mean(data_main$age, na.rm = T),2)
round(sd(data_main$age, na.rm = T),2)
```

#### Sex 
Gender
Categories: Woman; Genderqueer/Nonbinary; Man; None of the above; Prefer not to say; Other
```{r Sex}
table(data_main$sex) # in absolute numbers
round(table(data_main$sex)/sum(table(data_main$sex)),4)*100 # in percent
table(is.na(data_main$sex))
```

#### Relationship status
```{r}
table(data_main$relationship_status) # in absolute numbers
round(table(data_main$relationship_status)/sum(table(data_main$relationship_status)),3)*100 # in percent
table(is.na(data_main$relationship_status)) # NAs
```

#### Relationship length 
```{r}
table(data_main$relationship_length)
range(data_main$relationship_length, na.rm = T) # range
table(is.na(data_main$relationship_length)) # NAs
# histogram
hist(data_main$relationship_length)

# mean, sd
round(mean(data_main$relationship_length, na.rm = T),2)
round(sd(data_main$relationship_length, na.rm = T),2)
```

### Outcomes
#### Relationship options
```{r}
# mean
data_main %>%
  select(interest_hookups,
         interest_nonsexrel,
         interest_monrel,
         interest_nonmonrel,
         interest_altrel,
         interest_single,
         interest_parent) %>% 
  colMeans(na.rm = T) 

# standard deviation
round(apply(data_main %>% select(interest_hookups,
                                 interest_nonsexrel,
                                 interest_monrel,
                                 interest_nonmonrel,
                                 interest_altrel,
                                 interest_parent),2, sd, na.rm = T),2)

# check for NAs
table(is.na(data_main$interest_hookups))
table(is.na(data_main$interest_nonsexrel))
table(is.na(data_main$interest_monrel))
table(is.na(data_main$interest_nonmonrel))
table(is.na(data_main$interest_altrel))
table(is.na(data_main$interest_single))
table(is.na(data_main$interest_parent))
```

#### Partner preferences
```{r}
# mean
data_main %>% select(pref_imp_ks,
                     pref_imp_att,
                     pref_imp_fs,
                     pref_imp_ca,
                     pref_imp_ei,
                     pref_imp_sexually_experienced) %>% 
              colMeans(na.rm = T) 

# standard deviations
round(apply(data_main %>% select(pref_imp_ks,
                                 pref_imp_att,
                                 pref_imp_fs,
                                 pref_imp_ca,
                                 pref_imp_ei,
                                 pref_imp_sexually_experienced),2, sd, na.rm = T),2)

# check for NAs
table(is.na(data_main$pref_imp_ks))
table(is.na(data_main$pref_imp_att))
table(is.na(data_main$pref_imp_fs))
table(is.na(data_main$pref_imp_ca))
table(is.na(data_main$pref_imp_ei))
table(is.na(data_main$pref_imp_sexually_experienced))
```

#### Self-ratings
```{r}
# mean
data_main %>% select(self_ks,
                     self_att,
                     self_fs,
                     self_ca,
                     self_ei,
                     self_sexually_experienced) %>% 
              colMeans(na.rm = T) 


# standard deviation
round(apply(data_main %>% select(self_ks,
                                 self_att,
                                 self_fs,
                                 self_ca,
                                 self_ei,
                                 self_sexually_experienced),2, sd, na.rm = T),2)


# check for NAS
table(is.na(data_main$self_ks))
table(is.na(data_main$self_att))
table(is.na(data_main$self_fs))
table(is.na(data_main$self_ca))
table(is.na(data_main$self_ei))
table(is.na(data_main$self_sexually_experienced))
```


## Descriptives by sexual orientation
### Demographics
```{r}
# quick overview 
describeBy(data_main, data_main$sexual_orientation)

# split data set by sexual orientation
asexuals = data_main %>% filter(sexual_orientation == "Asexual")
heteros = data_main %>% filter(sexual_orientation == "Straight/Heterosexual")
```

#### Country
```{r}
# for asexuals
sort(table(asexuals$country), decreasing = T) # absolute numbers
sort(round(table(asexuals$country)/sum(table(asexuals$country)),4)*100, decreasing = T) # in percent
  
# for heterosexuals
sort(table(heteros$country), decreasing = T) # absolute numbers
sort(round(table(heteros$country)/sum(table(heteros$country)),4)*100, decreasing = T) # in percent


# check for NAs
table(is.na(asexuals$country))
table(is.na(heteros$country))
```

#### Language
```{r}
# for asexuals
sort(table(asexuals$language), decreasing = T) # absolute numbers
sort(round(table(asexuals$language)/sum(table(asexuals$language)),4)*100, decreasing = T) # in percent

# for heterosexuals
sort(table(heteros$language), decreasing = T) # absolute numbers
sort(round(table(heteros$language)/sum(table(heteros$language)),4)*100, decreasing = T) # in percent


# check for NAs
table(is.na(asexuals$language))
table(is.na(heteros$language))

```

#### Age
```{r}
# mean, sd, min, max
data_main %>%
  group_by(sexual_orientation) %>%
  summarise(mean(age, na.rm = T), 
            min(age, na.rm = T), 
            max(age, na.rm = T),
            sd(age, na.rm = T))


# Histogram by group 
ggplot(data_main, aes(x = age, fill = sexual_orientation)) + 
  geom_histogram()


# check NAs
data_main %>% 
  group_by(sexual_orientation) %>% 
  summarise(sum(is.na(age)))
```

#### Relationship status
```{r}
# for asexuals
sort(table(asexuals$relationship_status), decreasing = T) # absolute numbers
sort(round(table(asexuals$relationship_status)/sum(table(asexuals$relationship_status)),4)*100, decreasing = T) # in percent

# for heterosexuals
sort(table(heteros$relationship_status), decreasing = T) # absolute numbers
sort(round(table(heteros$relationship_status)/sum(table(heteros$relationship_status)),4)*100, decreasing = T) # in percent


# check for NAs
table(is.na(asexuals$relationship_status))
table(is.na(heteros$relationship_status))
```

#### Relationship length 
```{r}
# mean, sd, min, max
data_main %>%
  group_by(sexual_orientation) %>%
  summarise(mean(relationship_length, na.rm = T), 
            min(relationship_length, na.rm = T), 
            max(relationship_length, na.rm = T),
            sd(relationship_length, na.rm = T))


# Histogram by group 
ggplot(data_main, aes(x = relationship_length, fill = sexual_orientation)) + 
  geom_histogram()


# check NAs
data_main %>% 
  group_by(sexual_orientation) %>% 
  summarise(sum(is.na(relationship_length)))
```

### Outcomes by sexual orientation
#### Relationship options
```{r}
# mean
data_main %>% group_by(sexual_orientation) %>% 
  summarise_at(c("interest_hookups", "interest_nonsexrel", "interest_monrel",
                 "interest_nonmonrel","interest_altrel", "interest_single", 
                 "interest_parent"), mean, na.rm = T)


# standard deviation
data_main %>% group_by(sexual_orientation) %>% 
  summarise_at(c("interest_hookups", "interest_nonsexrel", "interest_monrel",
                 "interest_nonmonrel","interest_altrel", "interest_single", 
                 "interest_parent"), sd, na.rm = T)
# min
data_main %>% group_by(sexual_orientation) %>% 
  summarise_at(c("interest_hookups", "interest_nonsexrel", "interest_monrel",
                 "interest_nonmonrel","interest_altrel", "interest_single", 
                 "interest_parent"), min, na.rm = T)
# max
data_main %>% group_by(sexual_orientation) %>% 
  summarise_at(c("interest_hookups", "interest_nonsexrel", "interest_monrel",
                 "interest_nonmonrel","interest_altrel", "interest_single", 
                 "interest_parent"), max, na.rm = T)



# ausprobiert, mehere Variablen und gleihczeitg mehrere Funktionen
# mean = data_main %>% group_by(sexual_orientation) %>% 
#   summarise_at(vars(interest_hookups, interest_nonsexrel, interest_monrel,
#                  interest_nonmonrel,interest_altrel, interest_single, 
#                  interest_parent), funs(mean, min, max, sd), na.rm = T)


# check NAs
data_main %>% 
  group_by(sexual_orientation) %>% 
  summarise(sum(is.na(interest_hookups)))

data_main %>% 
  group_by(sexual_orientation) %>% 
  summarise(sum(is.na(interest_nonsexrel)))

data_main %>% 
  group_by(sexual_orientation) %>% 
  summarise(sum(is.na(interest_monrel)))

data_main %>% 
  group_by(sexual_orientation) %>% 
  summarise(sum(is.na(interest_nonmonrel)))

data_main %>% 
  group_by(sexual_orientation) %>% 
  summarise(sum(is.na(interest_altrel)))

data_main %>% 
  group_by(sexual_orientation) %>% 
  summarise(sum(is.na(interest_single)))

data_main %>% 
  group_by(sexual_orientation) %>% 
  summarise(sum(is.na(interest_parent)))


# Total number per variable by sexual orientation
data_main %>% group_by(sexual_orientation) %>% 
  summarise_at(c("interest_hookups", "interest_nonsexrel", "interest_monrel",
                 "interest_nonmonrel","interest_altrel", "interest_single", "interest_parent"), sum, na.rm = T)

```

#### Partner preferences
```{r}
# mean
data_main %>% group_by(sexual_orientation) %>% 
  summarise_at(c("pref_imp_ks",
                 "pref_imp_att",
                 "pref_imp_fs",
                 "pref_imp_ca",
                 "pref_imp_ei",
                 "pref_imp_sexually_experienced"), mean, na.rm = T)


# standard deviation
data_main %>% group_by(sexual_orientation) %>% 
  summarise_at(c("pref_imp_ks",
                 "pref_imp_att",
                 "pref_imp_fs",
                 "pref_imp_ca",
                 "pref_imp_ei",
                 "pref_imp_sexually_experienced"), sd, na.rm = T)
# min
data_main %>% group_by(sexual_orientation) %>% 
  summarise_at(c("pref_imp_ks",
                 "pref_imp_att",
                 "pref_imp_fs",
                 "pref_imp_ca",
                 "pref_imp_ei",
                 "pref_imp_sexually_experienced"), min, na.rm = T)
# max
data_main %>% group_by(sexual_orientation) %>% 
  summarise_at(c("pref_imp_ks",
                 "pref_imp_att",
                 "pref_imp_fs",
                 "pref_imp_ca",
                 "pref_imp_ei",
                 "pref_imp_sexually_experienced"), max, na.rm = T)


# check NAs
data_main %>% 
  group_by(sexual_orientation) %>% 
  summarise(sum(is.na(pref_imp_ks)))

data_main %>% 
  group_by(sexual_orientation) %>% 
  summarise(sum(is.na(pref_imp_att)))

data_main %>% 
  group_by(sexual_orientation) %>% 
  summarise(sum(is.na(pref_imp_fs)))

data_main %>% 
  group_by(sexual_orientation) %>% 
  summarise(sum(is.na(pref_imp_ca)))

data_main %>% 
  group_by(sexual_orientation) %>% 
  summarise(sum(is.na(pref_imp_ei)))

data_main %>% 
  group_by(sexual_orientation) %>% 
  summarise(sum(is.na(pref_imp_sexually_experienced)))




# Total number per variable by sexual orientation
data_main %>% group_by(sexual_orientation) %>% 
  summarise_at(c("pref_imp_ks",
                 "pref_imp_att",
                 "pref_imp_fs",
                 "pref_imp_ca",
                 "pref_imp_ei",
                 "pref_imp_sexually_experienced"), sum, na.rm = T)

```

#### Self ratings
```{r}
# mean
data_main %>% group_by(sexual_orientation) %>% 
  summarise_at(c("self_ks",
                 "self_att",
                 "self_fs",
                 "self_ca",
                 "self_ei",
                 "self_sexually_experienced"), mean, na.rm = T)


# standard deviation
data_main %>% group_by(sexual_orientation) %>% 
  summarise_at(c("self_ks",
                 "self_att",
                 "self_fs",
                 "self_ca",
                 "self_ei",
                 "self_sexually_experienced"), sd, na.rm = T)
# min
data_main %>% group_by(sexual_orientation) %>% 
  summarise_at(c("self_ks",
                 "self_att",
                 "self_fs",
                 "self_ca",
                 "self_ei",
                 "self_sexually_experienced"), min, na.rm = T)
# max
data_main %>% group_by(sexual_orientation) %>% 
  summarise_at(c("self_ks",
                 "self_att",
                 "self_fs",
                 "self_ca",
                 "self_ei",
                 "self_sexually_experienced"), max, na.rm = T)

# check NAs
data_main %>% 
  group_by(sexual_orientation) %>% 
  summarise(sum(is.na(self_ks)))

data_main %>% 
  group_by(sexual_orientation) %>% 
  summarise(sum(is.na(self_att)))

data_main %>% 
  group_by(sexual_orientation) %>% 
  summarise(sum(is.na(self_fs)))

data_main %>% 
  group_by(sexual_orientation) %>% 
  summarise(sum(is.na(self_ca)))

data_main %>% 
  group_by(sexual_orientation) %>% 
  summarise(sum(is.na(self_ei)))

data_main %>% 
  group_by(sexual_orientation) %>% 
  summarise(sum(is.na(self_sexually_experienced)))


# Total number per variable by sexual orientation
data_main %>% group_by(sexual_orientation) %>% 
  summarise_at(c("self_ks",
                 "self_att",
                 "self_fs",
                 "self_ca",
                 "self_ei",
                 "self_sexually_experienced"), sum, na.rm = T)

```

## Zero-order correlations between all outcomes
```{r}
correlations_all_outcomes = data_main %>%
  select(interest_hookups,
         interest_nonsexrel,
         interest_monrel,
         interest_nonmonrel,
         interest_altrel,
         interest_single,
         interest_parent,
         pref_imp_ks,
         pref_imp_att,
         pref_imp_fs,
         pref_imp_ca,
         pref_imp_ei,
         pref_imp_sexually_experienced,
         self_ks,
         self_att,
         self_fs,
         self_ca,
         self_ei,
         self_sexually_experienced)


correlations_table_all_outcomes = apa.cor.table(correlations_all_outcomes, filename = "Table_Intercorrelations.doc", table.number = 1)

correlations_table_all_outcomes
```
