==> 08_analyses.Rmd <==
---
title: "Analyses"
output: 
  html_document:
    toc: true
    toc_depth: 6
    toc_float: true
    code_folding: 'show'
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(warnings = FALSE)
```

## Library
```{r}
library(dplyr)
library(ggplot2)
library(psych)
library(effsize)
library(gridExtra)  
```


## Load data
```{r}
data_rel = read.csv(file = "data/data_rel.csv")[,-1]
data_prefs = read.csv(file = "data/data_prefs.csv")[,-1]
data_self = read.csv(file = "data/data_self.csv")[,-1]
```

## Sort data by subclass number within sexual orientation
```{r}
data_rel = data_rel %>% 
        group_by(sexual_orientation) %>%
        arrange(subclass, .by_group = TRUE)

data_prefs = data_prefs %>% 
        group_by(sexual_orientation) %>%
        arrange(subclass, .by_group = TRUE)

data_self = data_self %>% 
        group_by(sexual_orientation) %>%
        arrange(subclass, .by_group = TRUE)
```

## 1. Relationship options
### Interest non-sexual relationships
```{r}
t.test(interest_nonsexrel ~ sexual_orientation, 
       data = data_rel,
       alternative = "greater",
       paired = TRUE)

# effect size
effsize::cohen.d(interest_nonsexrel ~ sexual_orientation, data = data_rel, paired = T)
```

### Interest sexual relationships
```{r}
t.test(interest_hookups ~ sexual_orientation, 
       data = data_rel,
       alternative = "less",
       paired = TRUE)


# effect size
effsize::cohen.d(interest_hookups ~ sexual_orientation, data = data_rel, paired = T)
```

### Interest non-monogamous relationships
```{r}
t.test(interest_nonmonrel ~ sexual_orientation, 
       data = data_rel,
       alternative = "greater",
       paired = TRUE)


# effect size
effsize::cohen.d(interest_nonmonrel ~ sexual_orientation, data = data_rel, paired = T)
```

### Interest alternativ committed relationships
```{r}
t.test(interest_altrel ~ sexual_orientation, 
       data = data_rel,
       alternative = "greater",
       paired = TRUE)


# effect size
effsize::cohen.d(interest_altrel ~ sexual_orientation, data = data_rel, paired = T)
```

### Interest single
```{r}
t.test(interest_single ~ sexual_orientation, 
       data = data_rel,
       alternative = "greater",
       paired = TRUE)


# effect size
effsize::cohen.d(interest_single ~ sexual_orientation, data = data_rel, paired = T)
```

### Exploratory analyses
#### Interest monogamous relationships
```{r}
t.test(interest_monrel ~ sexual_orientation, 
       data = data_rel,
       alternative = "two.sided",
       paired = TRUE)


# effect size
effsize::cohen.d(interest_monrel ~ sexual_orientation, data = data_rel, paired = T)
```

#### Interest parent 
```{r}
t.test(interest_parent ~ sexual_orientation, 
       data = data_rel,
       alternative = "two.sided",
       paired = TRUE)


# effect size
effsize::cohen.d(interest_parent ~ sexual_orientation, data = data_rel, paired = T)
```


## 2. Partner preferences
### Confident-assertive
```{r}
t.test(pref_imp_ca ~ sexual_orientation, 
       data = data_prefs,
       alternative = "less",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_ca ~ sexual_orientation, data = data_prefs, paired = T)
```

### Attractive
```{r}
t.test(pref_imp_att ~ sexual_orientation, 
       data = data_prefs,
       alternative = "less",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_att ~ sexual_orientation, data = data_prefs, paired = T)
```

### Sexually experienced
```{r}
t.test(pref_imp_sexually_experienced ~ sexual_orientation, 
       data = data_prefs,
       alternative = "less",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_sexually_experienced ~ sexual_orientation, data = data_prefs, paired = T)
```

### Exploratory analyses
#### Kind-supportive
```{r}
t.test(pref_imp_ks ~ sexual_orientation, 
       data = data_prefs,
       alternative = "two.sided",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_ks ~ sexual_orientation, data = data_prefs, paired = T)
```


#### Financially secure-successful
```{r}
t.test(pref_imp_fs ~ sexual_orientation, 
       data = data_prefs,
       alternative = "two.sided",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_fs ~ sexual_orientation, data = data_prefs, paired = T)
```


#### Educated-intelligent
```{r}
t.test(pref_imp_ei ~ sexual_orientation, 
       data = data_prefs,
       alternative = "two.sided",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_ei ~ sexual_orientation, data = data_prefs, paired = T)
```


## 3. Self-ratings
### Confident-assertive
```{r}
t.test(self_ca ~ sexual_orientation, 
       data = data_self,
       alternative = "less",
       paired = TRUE)


# effect size
effsize::cohen.d(self_ca ~ sexual_orientation, data = data_self, paired = T)
```

### Attractive
```{r}
t.test(self_att ~ sexual_orientation, 
       data = data_self,
       alternative = "less",
       paired = TRUE)


# effect size
effsize::cohen.d(self_att ~ sexual_orientation, data = data_self, paired = T)
```

### Sexually experienced
```{r}
t.test(self_sexually_experienced ~ sexual_orientation, 
       data = data_self,
       alternative = "less",
       paired = TRUE)


# effect size
effsize::cohen.d(self_sexually_experienced ~ sexual_orientation, data = data_self, paired = T)
```

### Exploratory analyses
#### Kind-supportive
```{r}
t.test(self_ks ~ sexual_orientation, 
       data = data_self,
       alternative = "two.sided",
       paired = TRUE)


# effect size
effsize::cohen.d(self_ks ~ sexual_orientation, data = data_self, paired = T)
```


#### Financially secure-successful
```{r}
t.test(self_fs ~ sexual_orientation, 
       data = data_self,
       alternative = "two.sided",
       paired = TRUE)


# effect size
effsize::cohen.d(self_fs ~ sexual_orientation, data = data_self, paired = T)
```

#### Educated-intelligent
```{r}
t.test(self_ei ~ sexual_orientation, 
       data = data_self,
       alternative = "two.sided",
       paired = TRUE)


# effect size
effsize::cohen.d(self_ei ~ sexual_orientation, data = data_self, paired = T)
```


## 4. Item-level analyses for partner preferences and self-ratings
### Add variabes to data set
```{r}
# load data_main
data_main = read.csv(file = "data/data_main.csv")[,-1]


# for partner preferences:
# select item level variables for partner preferences
data_main_pref_items = data_main %>% select(id, pref_imp_confident, pref_imp_assertive, 
                                           pref_imp_attractive_body, pref_imp_attractive_face,
                                           pref_imp_kind, pref_imp_supportive,
                                           pref_imp_educated, pref_imp_intelligence,
                                           pref_imp_financially_secure, pref_imp_successful_ambitous)

# create new data set by adding item level variables to data_prefs
data_prefs_itemlevel = left_join(data_prefs, data_main_pref_items, by = "id")


# for self-ratings
# select item level variables for partner preferences
data_main_self_items = data_main %>% select(id, self_confident, self_assertive, 
                                           self_attractive_body, self_attractive_face,
                                           self_kind, self_supportive,
                                           self_educated, self_intelligence,
                                           self_financially_secure, self_successful_ambitous)

# create new data set by adding item level variables to data_prefs
data_self_itemlevel = left_join(data_self, data_main_self_items, by = "id")
```

### Overview
### For partner preferences
```{r}
# quick overview by sexual orientation
describe.by(data_prefs_itemlevel, data_prefs_itemlevel$sexual_orientation)

# mean
data_prefs_itemlevel %>% group_by(sexual_orientation) %>% 
  summarise_at(c("pref_imp_confident", "pref_imp_assertive",
                 "pref_imp_attractive_body", "pref_imp_attractive_face",
                 "pref_imp_kind","pref_imp_supportive",
                 "pref_imp_financially_secure", "pref_imp_successful_ambitous", 
                 "pref_imp_educated", "pref_imp_intelligence",
                 "pref_imp_sexually_experienced"), mean, na.rm = T)


# standard deviation
data_prefs_itemlevel %>% group_by(sexual_orientation) %>% 
  summarise_at(c("pref_imp_confident", "pref_imp_assertive",
                 "pref_imp_attractive_body", "pref_imp_attractive_face",
                 "pref_imp_kind","pref_imp_supportive",
                 "pref_imp_financially_secure", "pref_imp_successful_ambitous", 
                 "pref_imp_educated", "pref_imp_intelligence",
                 "pref_imp_sexually_experienced"), sd, na.rm = T)
# min
data_prefs_itemlevel %>% group_by(sexual_orientation) %>% 
  summarise_at(c("pref_imp_confident", "pref_imp_assertive",
                 "pref_imp_attractive_body", "pref_imp_attractive_face",
                 "pref_imp_kind","pref_imp_supportive",
                 "pref_imp_financially_secure", "pref_imp_successful_ambitous", 
                 "pref_imp_educated", "pref_imp_intelligence",
                 "pref_imp_sexually_experienced"), min, na.rm = T)
# max
data_prefs_itemlevel %>% group_by(sexual_orientation) %>% 
  summarise_at(c("pref_imp_confident", "pref_imp_assertive",
                 "pref_imp_attractive_body", "pref_imp_attractive_face",
                 "pref_imp_kind","pref_imp_supportive",
                 "pref_imp_financially_secure", "pref_imp_successful_ambitous", 
                 "pref_imp_educated", "pref_imp_intelligence",
                 "pref_imp_sexually_experienced"), max, na.rm = T)
```

### For self-ratings
```{r}
# quick overview by sexual orientation
describe.by(data_self_itemlevel, data_self_itemlevel$sexual_orientation)

# mean
data_self_itemlevel %>% group_by(sexual_orientation) %>% 
  summarise_at(c("self_confident", "self_assertive",
                 "self_attractive_body", "self_attractive_face",
                 "self_kind","self_supportive",
                 "self_financially_secure", "self_successful_ambitous", 
                 "self_educated", "self_intelligence",
                 "self_sexually_experienced"), mean, na.rm = T)


# standard deviation
data_self_itemlevel %>% group_by(sexual_orientation) %>% 
  summarise_at(c("self_confident", "self_assertive",
                 "self_attractive_body", "self_attractive_face",
                 "self_kind","self_supportive",
                 "self_financially_secure", "self_successful_ambitous", 
                 "self_educated", "self_intelligence",
                 "self_sexually_experienced"), sd, na.rm = T)
# min
data_self_itemlevel %>% group_by(sexual_orientation) %>% 
  summarise_at(c("self_confident", "self_assertive",
                 "self_attractive_body", "self_attractive_face",
                 "self_kind","self_supportive",
                 "self_financially_secure", "self_successful_ambitous", 
                 "self_educated", "self_intelligence",
                 "self_sexually_experienced"), min, na.rm = T)
# max
data_self_itemlevel %>% group_by(sexual_orientation) %>% 
  summarise_at(c("self_confident", "self_assertive",
                 "self_attractive_body", "self_attractive_face",
                 "self_kind","self_supportive",
                 "self_financially_secure", "self_successful_ambitous", 
                 "self_educated", "self_intelligence",
                 "self_sexually_experienced"), max, na.rm = T)
```

### Creates subsamples for sexual orientation
```{r}
# for partner preferences
asex_data_prefs_itemlevel = data_prefs_itemlevel %>% filter(sexual_orientation == "Asexual") %>% arrange(subclass)
hetero_data_prefs_itemlevel = data_prefs_itemlevel %>% filter(sexual_orientation == "Straight/Heterosexual") %>% arrange(subclass)

# for self-ratings
asex_data_self_itemlevel = data_self_itemlevel %>% filter(sexual_orientation == "Asexual") %>% arrange(subclass)
hetero_data_self_itemlevel = data_self_itemlevel %>% filter(sexual_orientation == "Straight/Heterosexual") %>% arrange(subclass)

```

### 4.1 Partner preferences
#### Confident-assertive
```{r}
# compare distributions
prefs_confident = ggplot(data_prefs_itemlevel, aes(x = pref_imp_confident, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge", binwidth = 0.5) +
  ylim(0,150) +
  ggtitle("Confident")

prefs_assertive = ggplot(data_prefs_itemlevel, aes(x = pref_imp_assertive, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge",  binwidth = 0.5) +
  ylim(0,150) +
  ggtitle("Assertive")

grid.arrange(prefs_confident, prefs_assertive, ncol = 2)  
```

##### Confident
```{r}
# mean, sd, min, max
data_prefs_itemlevel %>%
  group_by(sexual_orientation) %>%
  summarise(mean(pref_imp_confident, na.rm = T), 
            sd(pref_imp_confident, na.rm = T), 
            min(pref_imp_confident, na.rm = T),
            max(pref_imp_confident, na.rm = T))

# compute t-test
t.test(asex_data_prefs_itemlevel$pref_imp_confident, hetero_data_prefs_itemlevel$pref_imp_confident,
       alternative = "less",
       paired = TRUE)


# effect size
effsize::cohen.d(asex_data_prefs_itemlevel$pref_imp_confident, hetero_data_prefs_itemlevel$pref_imp_confident,
       alternative = "less",
       paired = TRUE)
```

##### Assertive
```{r}
# mean, sd, min, max
data_prefs_itemlevel %>%
  group_by(sexual_orientation) %>%
  summarise(mean(pref_imp_assertive, na.rm = T), 
            sd(pref_imp_assertive, na.rm = T), 
            min(pref_imp_assertive, na.rm = T),
            max(pref_imp_assertive, na.rm = T))


t.test(asex_data_prefs_itemlevel$pref_imp_assertive, hetero_data_prefs_itemlevel$pref_imp_assertive, 
       alternative = "less",
       paired = TRUE) 

# effect size
effsize::cohen.d(asex_data_prefs_itemlevel$pref_imp_assertive, hetero_data_prefs_itemlevel$pref_imp_assertive, 
       alternative = "less",
       paired = TRUE,
       na.rm = T)
```


#### Attractive
```{r}
# compare distributions
prefs_attractive_body = ggplot(data_prefs_itemlevel, aes(x = pref_imp_attractive_body, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge", binwidth = 0.5) +
  ylim(0,150) +
  ggtitle("Attractive body") 

prefs_attractive_face = ggplot(data_prefs_itemlevel, aes(x = pref_imp_attractive_face, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge", binwidth = 0.5) +
  ylim(0,150) +
  ggtitle("Attractive face")

grid.arrange(prefs_attractive_body, prefs_attractive_face, ncol = 2) 

# correlation between attractive body and attractive face
cor.test(data_prefs_itemlevel$pref_imp_attractive_body, data_prefs_itemlevel$pref_imp_attractive_face, method=c("pearson"))
```
Correlation between attractive face and body r = .69.



##### Attractive body
```{r}
# mean, sd, min, max
data_prefs_itemlevel %>%
  group_by(sexual_orientation) %>%
  summarise(mean(pref_imp_attractive_body, na.rm = T), 
            sd(pref_imp_attractive_body, na.rm = T), 
            min(pref_imp_attractive_body, na.rm = T),
            max(pref_imp_attractive_body, na.rm = T))

# compute t-test
t.test(asex_data_prefs_itemlevel$pref_imp_attractive_body, hetero_data_prefs_itemlevel$pref_imp_attractive_body,
       alternative = "less",
       paired = TRUE)


# effect size
effsize::cohen.d(asex_data_prefs_itemlevel$pref_imp_attractive_body, hetero_data_prefs_itemlevel$pref_imp_attractive_body,
       alternative = "less",
       paired = TRUE,
       na.rm = T)
```

##### Attractive face
```{r}
# mean, sd, min, max
data_prefs_itemlevel %>%
  group_by(sexual_orientation) %>%
  summarise(mean(pref_imp_attractive_face, na.rm = T), 
            sd(pref_imp_attractive_face, na.rm = T), 
            min(pref_imp_attractive_face, na.rm = T),
            max(pref_imp_attractive_face, na.rm = T))

# compute t-test
t.test(asex_data_prefs_itemlevel$pref_imp_attractive_face, hetero_data_prefs_itemlevel$pref_imp_attractive_face,
       data = data_prefs_itemlevel,
       alternative = "less",
       paired = TRUE)


# effect size
effsize::cohen.d(asex_data_prefs_itemlevel$pref_imp_attractive_face, hetero_data_prefs_itemlevel$pref_imp_attractive_face,
       data = data_prefs_itemlevel,
       alternative = "less",
       paired = TRUE,
       na.rm = T)
```


#### Kind-supportive
```{r}
# compare distributions
prefs_kind = ggplot(data_prefs_itemlevel, aes(x = pref_imp_kind, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge") +
  ylim(0,300) +
  ggtitle("Kind") 

prefs_supportive = ggplot(data_prefs_itemlevel, aes(x = pref_imp_supportive, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge") +
  ylim(0,300) +
  ggtitle("Supportive")

grid.arrange(prefs_kind, prefs_supportive, ncol = 2)  
```

##### Kind
```{r}
# mean, sd, min, max
data_prefs_itemlevel %>%
  group_by(sexual_orientation) %>%
  summarise(mean(pref_imp_kind, na.rm = T), 
            sd(pref_imp_kind, na.rm = T), 
            min(pref_imp_kind, na.rm = T),
            max(pref_imp_kind, na.rm = T))

# compute t-test
t.test(asex_data_prefs_itemlevel$pref_imp_kind, hetero_data_prefs_itemlevel$pref_imp_kind,
       alternative = "two.sided",
       paired = T)


# effect size
effsize::cohen.d(asex_data_prefs_itemlevel$pref_imp_kind, hetero_data_prefs_itemlevel$pref_imp_kind,
       alternative = "two.sided",
       paired = T,
       na.rm = T)
```


##### Supportive
```{r}
# mean, sd, min, max
data_prefs_itemlevel %>%
  group_by(sexual_orientation) %>%
  summarise(mean(pref_imp_supportive, na.rm = T), 
            sd(pref_imp_supportive, na.rm = T), 
            min(pref_imp_supportive, na.rm = T),
            max(pref_imp_supportive, na.rm = T))

# compute t-test
t.test(asex_data_prefs_itemlevel$pref_imp_supportive, hetero_data_prefs_itemlevel$pref_imp_supportive,
       alternative = "two.sided",
       paired = T)


# effect size
effsize::cohen.d(asex_data_prefs_itemlevel$pref_imp_supportive, hetero_data_prefs_itemlevel$pref_imp_supportive,
       alternative = "two.sided",
       paired = T,
       na.rm = T)
```


#### Financially secure-successful
```{r}
# compare distributions
prefs_financially_secure = ggplot(data_prefs_itemlevel, aes(x = pref_imp_financially_secure, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge") +
  ylim(0,200) +
  ggtitle("Financially secure") 

prefs_successful_ambitious = ggplot(data_prefs_itemlevel, aes(x = pref_imp_successful_ambitous, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge") +
  ylim(0,200) +
  ggtitle("Successful ambitious")

grid.arrange(prefs_financially_secure, prefs_successful_ambitious, ncol = 2)  
```

##### Financially secure
```{r}
# mean, sd, min, max
data_prefs_itemlevel %>%
  group_by(sexual_orientation) %>%
  summarise(mean(pref_imp_financially_secure, na.rm = T), 
            sd(pref_imp_financially_secure, na.rm = T), 
            min(pref_imp_financially_secure, na.rm = T),
            max(pref_imp_financially_secure, na.rm = T))

# compute t-test
t.test(asex_data_prefs_itemlevel$pref_imp_financially_secure,hetero_data_prefs_itemlevel$pref_imp_financially_secure,
       data = data_prefs_itemlevel,
       alternative = "two.sided",
       paired = T)


# effect size
effsize::cohen.d(asex_data_prefs_itemlevel$pref_imp_financially_secure,hetero_data_prefs_itemlevel$pref_imp_financially_secure,
       data = data_prefs_itemlevel,
       alternative = "two.sided",
       paired = T,
       na.rm = T)
```


##### Successful ambitious
```{r}
# mean, sd, min, max
data_prefs_itemlevel %>%
  group_by(sexual_orientation) %>%
  summarise(mean(pref_imp_successful_ambitous, na.rm = T), 
            sd(pref_imp_successful_ambitous, na.rm = T), 
            min(pref_imp_successful_ambitous, na.rm = T),
            max(pref_imp_successful_ambitous, na.rm = T))

# compute t-test
t.test(asex_data_prefs_itemlevel$pref_imp_successful_ambitous, hetero_data_prefs_itemlevel$pref_imp_successful_ambitous,
       alternative = "two.sided",
       paired = T)


# effect size
effsize::cohen.d(asex_data_prefs_itemlevel$pref_imp_successful_ambitous, hetero_data_prefs_itemlevel$pref_imp_successful_ambitous,
       alternative = "two.sided",
       paired = T,
       na.rm = T)
```

#### Educated-intelligent
```{r}
# compare distributions
prefs_educated = ggplot(data_prefs_itemlevel, aes(x = pref_imp_educated, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge") +
  ylim(0,200) +
  ggtitle("Edcuated") 

prefs_intelligent = ggplot(data_prefs_itemlevel, aes(x = pref_imp_intelligence, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge") +
  ylim(0,200) +
  ggtitle("Intelligent")

grid.arrange(prefs_educated, prefs_intelligent, ncol = 2)  
```

##### Educated
```{r}
# mean, sd, min, max
data_prefs_itemlevel %>%
  group_by(sexual_orientation) %>%
  summarise(mean(pref_imp_educated, na.rm = T), 
            sd(pref_imp_educated, na.rm = T), 
            min(pref_imp_educated, na.rm = T),
            max(pref_imp_educated, na.rm = T))

# compute t-test
t.test(asex_data_prefs_itemlevel$pref_imp_educated,hetero_data_prefs_itemlevel$pref_imp_educated,
       alternative = "two.sided",
       paired = T)


# effect size
effsize::cohen.d(asex_data_prefs_itemlevel$pref_imp_educated,hetero_data_prefs_itemlevel$pref_imp_educated,
       alternative = "two.sided",
       paired = T,
       na.rm = T)
```

##### Intelligent
```{r}
# mean, sd, min, max
data_prefs_itemlevel %>%
  group_by(sexual_orientation) %>%
  summarise(mean(pref_imp_intelligence, na.rm = T), 
            sd(pref_imp_intelligence, na.rm = T), 
            min(pref_imp_intelligence, na.rm = T),
            max(pref_imp_intelligence, na.rm = T))

# compute t-test
t.test(asex_data_prefs_itemlevel$pref_imp_intelligence, hetero_data_prefs_itemlevel$pref_imp_intelligence,
       alternative = "two.sided",
       paired = T)


# effect size
effsize::cohen.d(asex_data_prefs_itemlevel$pref_imp_intelligence, hetero_data_prefs_itemlevel$pref_imp_intelligence,
       alternative = "two.sided",
       paired = T,
       na.rm = T)
```


### 4.2 Self-ratings
#### Confident-assertive
```{r}
# compute t-test
# compare distributions
self_confident = ggplot(data_self_itemlevel, aes(x = self_confident, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge", binwidth = 0.5) +
  ylim(0,150) +
  ggtitle("Confident")

self_assertive = ggplot(data_self_itemlevel, aes(x = self_assertive, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge", binwidth = 0.5) +
  ylim(0,150) +
  ggtitle("Assertive")

grid.arrange(self_confident, self_assertive, ncol = 2)  
```

##### Confident
```{r}
# mean, sd, min, max
data_self_itemlevel %>%
  group_by(sexual_orientation) %>%
  summarise(mean(self_confident, na.rm = T), 
            sd(self_confident, na.rm = T), 
            min(self_confident, na.rm = T),
            max(self_confident, na.rm = T))

# compute t-test
t.test(asex_data_self_itemlevel$self_confident, hetero_data_self_itemlevel$self_confident,
       alternative = "less",
       paired = T)


# effect size
effsize::cohen.d(asex_data_self_itemlevel$self_confident, hetero_data_self_itemlevel$self_confident,
       alternative = "less",
       paired = T,
       na.rm = T)
```


##### Assertive
```{r}
# mean, sd, min, max
data_self_itemlevel %>%
  group_by(sexual_orientation) %>%
  summarise(mean(self_assertive, na.rm = T), 
            sd(self_assertive, na.rm = T), 
            min(self_assertive, na.rm = T),
            max(self_assertive, na.rm = T))

# compute t-test
t.test(asex_data_self_itemlevel$self_assertive, hetero_data_self_itemlevel$self_assertive,
       alternative = "less",
       paired = T) 

# effect size
effsize::cohen.d(asex_data_self_itemlevel$self_assertive, hetero_data_self_itemlevel$self_assertive,
       alternative = "less",
       paired = T,
       na.rm = T)
```


#### Attractive
```{r}
# compare distributions
self_attractive_body = ggplot(data_self_itemlevel, aes(x = self_attractive_body, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge", binwidth = 0.5) +
  ylim(0,150) +
  ggtitle("Attractive body") 

self_attractive_face = ggplot(data_self_itemlevel, aes(x = self_attractive_face, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge", binwidth = 0.5) +
  ylim(0,150) +
  ggtitle("Attractive face")

grid.arrange(self_attractive_body, self_attractive_face, ncol = 2)  

# correlation between attractive body and attractive face
cor.test(data_self_itemlevel$self_attractive_body, data_self_itemlevel$self_attractive_face, method=c("pearson"))
```
Correlation between attractive body and face r = 0.66. 


##### Attractive body
```{r}
# mean, sd, min, max
data_self_itemlevel %>%
  group_by(sexual_orientation) %>%
  summarise(mean(self_attractive_body, na.rm = T), 
            sd(self_attractive_body, na.rm = T), 
            min(self_attractive_body, na.rm = T),
            max(self_attractive_body, na.rm = T))

# compute t-test
t.test(asex_data_self_itemlevel$self_attractive_body, hetero_data_self_itemlevel$self_attractive_body,
       alternative = "less",
       paired = T)


# effect size
effsize::cohen.d(asex_data_self_itemlevel$self_attractive_body, hetero_data_self_itemlevel$self_attractive_body,
       alternative = "less",
       paired = T,
       na.rm = T)
```


##### Attractive face
```{r}
# mean, sd, min, max
data_self_itemlevel %>%
  group_by(sexual_orientation) %>%
  summarise(mean(self_attractive_face, na.rm = T), 
            sd(self_attractive_face, na.rm = T), 
            min(self_attractive_face, na.rm = T),
            max(self_attractive_face, na.rm = T))

# compute t-test
t.test(asex_data_self_itemlevel$self_attractive_face, hetero_data_self_itemlevel$self_attractive_face,
       alternative = "less",
       paired = T)


# effect size
effsize::cohen.d(asex_data_self_itemlevel$self_attractive_face, hetero_data_self_itemlevel$self_attractive_face,
       alternative = "less",
       paired = T,
       na.rm = T)
```

#### Kind-supportive
```{r}
# compare distributions
self_kind = ggplot(data_self_itemlevel, aes(x = self_kind, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge") +
  ylim(0,300) +
  ggtitle("Kind") 

self_supportive = ggplot(data_self_itemlevel, aes(x = self_supportive, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge") +
  ylim(0,300) +
  ggtitle("Supportive")

grid.arrange(self_kind, self_supportive, ncol = 2)  
```

##### Kind
```{r}
# mean, sd, min, max
data_self_itemlevel %>%
  group_by(sexual_orientation) %>%
  summarise(mean(self_kind, na.rm = T), 
            sd(self_kind, na.rm = T), 
            min(self_kind, na.rm = T),
            max(self_kind, na.rm = T))

# compute t-test
t.test(asex_data_self_itemlevel$self_kind, hetero_data_self_itemlevel$self_kind,
       alternative = "two.sided",
       paired = T)


# effect size
effsize::cohen.d(asex_data_self_itemlevel$self_kind, hetero_data_self_itemlevel$self_kind,
       alternative = "two.sided",
       paired = T,
       na.rm = T)
```

##### Supportive
```{r}
# mean, sd, min, max
data_self_itemlevel %>%
  group_by(sexual_orientation) %>%
  summarise(mean(self_supportive, na.rm = T), 
            sd(self_supportive, na.rm = T), 
            min(self_supportive, na.rm = T),
            max(self_supportive, na.rm = T))

# compute t-test
t.test(asex_data_self_itemlevel$self_supportive,hetero_data_self_itemlevel$self_supportive,
       alternative = "two.sided",
       paired = T)


# effect size
effsize::cohen.d(asex_data_self_itemlevel$self_supportive,hetero_data_self_itemlevel$self_supportive,
       alternative = "two.sided",
       paired = T,
       na.rm = T)
```

#### Financially secure-successful
```{r}
# compare distributions
self_financially_secure = ggplot(data_self_itemlevel, aes(x = self_financially_secure, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge") +
  ylim(0,200) +
  ggtitle("Financially secure") 

self_successful_ambitious = ggplot(data_self_itemlevel, aes(x = self_successful_ambitous, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge") +
  ylim(0,200) +
  ggtitle("Successful ambitious")

grid.arrange(self_financially_secure, self_successful_ambitious, ncol = 2)  
```

##### Financially secure
```{r}
# mean, sd, min, max
data_self_itemlevel %>%
  group_by(sexual_orientation) %>%
  summarise(mean(self_financially_secure, na.rm = T), 
            sd(self_financially_secure, na.rm = T), 
            min(self_financially_secure, na.rm = T),
            max(self_financially_secure, na.rm = T))

# compute t-test
t.test(asex_data_self_itemlevel$self_financially_secure,hetero_data_self_itemlevel$self_financially_secure,
       alternative = "two.sided",
       paired = T)


# effect size
effsize::cohen.d(asex_data_self_itemlevel$self_financially_secure,hetero_data_self_itemlevel$self_financially_secure,
       alternative = "two.sided",
       paired = T,
       na.rm = T)
```


##### Successful ambitious
```{r}
# mean, sd, min, max
data_self_itemlevel %>%
  group_by(sexual_orientation) %>%
  summarise(mean(self_successful_ambitous, na.rm = T), 
            sd(self_successful_ambitous, na.rm = T), 
            min(self_successful_ambitous, na.rm = T),
            max(self_successful_ambitous, na.rm = T))

# compute t-test
t.test(asex_data_self_itemlevel$self_successful_ambitous,hetero_data_self_itemlevel$self_successful_ambitous,
       alternative = "two.sided",
       paired = T)


# effect size
effsize::cohen.d(asex_data_self_itemlevel$self_successful_ambitous,hetero_data_self_itemlevel$self_successful_ambitous,
       alternative = "two.sided",
       paired = T,
       na.rm = T)
```


#### Educated-intelligent
```{r}
# compare distributions
self_educated = ggplot(data_self_itemlevel, aes(x = self_educated, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge") +
  ylim(0,200) +
  ggtitle("Edcuated") 

self_intelligent = ggplot(data_self_itemlevel, aes(x = self_intelligence, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge") +
  ylim(0,200) +
  ggtitle("Intelligent")

grid.arrange(self_educated, self_intelligent, ncol = 2)  
```

##### Educated
```{r}
# mean, sd, min, max
data_self_itemlevel %>%
  group_by(sexual_orientation) %>%
  summarise(mean(self_educated, na.rm = T), 
            sd(self_educated, na.rm = T), 
            min(self_educated, na.rm = T),
            max(self_educated, na.rm = T))

# compute t-test
t.test(asex_data_self_itemlevel$self_educated,hetero_data_self_itemlevel$self_educated,
       alternative = "two.sided",
       paired = T)


# effect size
effsize::cohen.d(asex_data_self_itemlevel$self_educated,hetero_data_self_itemlevel$self_educated,
       alternative = "two.sided",
       paired = T,
       na.rm = T)
```


##### Intelligent
```{r}
# mean, sd, min, max
data_self_itemlevel %>%
  group_by(sexual_orientation) %>%
  summarise(mean(self_intelligence, na.rm = T), 
            sd(self_intelligence, na.rm = T), 
            min(self_intelligence, na.rm = T),
            max(self_intelligence, na.rm = T))

# compute t-test
t.test(asex_data_self_itemlevel$self_intelligence,hetero_data_self_itemlevel$self_intelligence,
       alternative = "two.sided",
       paired = T)


# effect size
effsize::cohen.d(asex_data_self_itemlevel$self_intelligence,hetero_data_self_itemlevel$self_intelligence,
       alternative = "two.sided",
       paired = T,
       na.rm = T)
```
