==> 09_comparison_matched_unmatched.Rmd <==
---
title: "Comparison Between Matched Asexuals, Matched Heterosexuals, and Unmatched Heterosexuals"
output: 
  html_document:
    toc: true
    toc_depth: 6
    toc_float: true
    code_folding: 'show'
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(warning = FALSE)
```

## Library
```{r}
library(dplyr)
library(ggplot2)
library(psych)
library(rcompanion) 
library(confintr)
```

## Load data
```{r}
# matched samples
data_rel = read.csv(file = "data/data_rel.csv")[,-1]
data_prefs = read.csv(file = "data/data_prefs.csv")[,-1]
data_self = read.csv(file = "data/data_self.csv")[,-1]

# unmatched heterosexual sample (includes only heterosexuals because during psm no asexual was discarded)
data_rel_unmatched = read.csv(file = "data/data_rel_unmatched.csv")[,-1]
data_prefs_unmatched = read.csv(file = "data/data_prefs_unmatched.csv")[,-1]
data_self_unmatched = read.csv(file = "data/data_self_unmatched.csv")[,-1]

# asexual sample 
asexuals_rel = data_rel %>% filter(sexual_orientation == "Asexual")
asexuals_prefs = data_prefs %>% filter(sexual_orientation == "Asexual")
asexuals_self = data_self %>% filter(sexual_orientation == "Asexual")

# matched heterosexual sample (also from partner preference psm sample)
heterosexuals_rel = data_rel %>%  filter(sexual_orientation == "Straight/Heterosexual")
heterosexuals_prefs = data_prefs %>%  filter(sexual_orientation == "Straight/Heterosexual")
heterosexuals_self = data_self %>%  filter(sexual_orientation == "Straight/Heterosexual")
```


## Create samples for each comparison:
```{r  }
# Asexuals v. matched heterosexuals: 
asex_v_hetero_rel = data_rel
asex_v_hetero_prefs = data_prefs
asex_v_hetero_self = data_self

# Asexuals v. unmatched heterosexuals:
asex_v_umhetero_rel = full_join(asexuals_rel, data_rel_unmatched)
asex_v_umhetero_prefs = full_join(asexuals_prefs, data_prefs_unmatched)
asex_v_umhetero_self = full_join(asexuals_self, data_self_unmatched)

# Matched heterosexuals v. unmatched heterosexuals:
# add variable matched (matched or unmatched) as grouping variable to data sets
data_rel_unmatched$matched = "unmatched"
heterosexuals_rel$matched = "matched"

data_prefs_unmatched$matched = "unmatched"
heterosexuals_prefs$matched = "matched"

data_self_unmatched$matched = "unmatched"
heterosexuals_self$matched = "matched"

# merge data sets
hetero_v_umhetero_rel = full_join(heterosexuals_rel, data_rel_unmatched)
hetero_v_umhetero_prefs = full_join(heterosexuals_prefs, data_prefs_unmatched)
hetero_v_umhetero_self = full_join(heterosexuals_self, data_self_unmatched)
```


## 0. Descriptives for unmatched Heterosexuals
[For asexuals and matched heterosexuals descriptives have been computed before - see 07_descriptives_psm.Rmd]
### Covariates (based on partner preference sample)
```{r}
# quick overview 
describeBy(data_prefs_unmatched)
```

#### Country
```{r}
sort(table(data_prefs_unmatched$country), decreasing = T) # absolute numbers
sort(round(table(data_prefs_unmatched$country)/sum(table(data_prefs_unmatched$country)),4)*100, decreasing = T) # in percent
```

#### Language
```{r}
sort(table(data_prefs_unmatched$language), decreasing = T) # absolute numbers
sort(round(table(data_prefs_unmatched$language)/sum(table(data_prefs_unmatched$language)),4)*100, decreasing = T) # in percent
```

#### Age
```{r}
# mean, sd, min, max
data_prefs_unmatched %>%
  summarise(mean(age, na.rm = T), 
            min(age, na.rm = T), 
            max(age, na.rm = T),
            sd(age, na.rm = T))


# Histogram 
ggplot(data_prefs_unmatched, aes(x = age, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge")
```

#### Relationship status
```{r}
sort(table(data_prefs_unmatched$relationship_status), decreasing = T) # absolute numbers
sort(round(table(data_prefs_unmatched$relationship_status)/sum(table(data_prefs_unmatched$relationship_status)),4)*100, decreasing = T) # in percent
```

#### Relationship length 
```{r}
# mean, sd, min, max
data_prefs_unmatched %>%
  summarise(mean(relationship_length, na.rm = T), 
            min(relationship_length, na.rm = T), 
            max(relationship_length, na.rm = T),
            sd(relationship_length, na.rm = T))


# Histogram by group 
ggplot(data_prefs, aes(x = relationship_length, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge")
```




## 1. Asexuals v. matched heterosexuals
### Sort matched data by subclass number within sexual_orientation (for dependent t-test order of observation matters)
```{r}
data_rel = data_rel %>% 
        group_by(sexual_orientation) %>%
        arrange(subclass, .by_group = TRUE)

data_prefs = data_prefs %>% 
        group_by(sexual_orientation) %>%
        arrange(subclass, .by_group = TRUE)

data_self = data_self %>% 
        group_by(sexual_orientation) %>%
        arrange(subclass, .by_group = TRUE)
```

### Covariates
#### Country
```{r}
# assumptions not met for chi square test: some cells contain less than 5 observations; therefore a Fisher's exact test is calculated
matrix_country = as.matrix(table(data_prefs$country, data_prefs$sexual_orientation), header=TRUE, row.names=1)
table(matrix_country < 5) # absolute number of cells containing less than 5 counts
nrow(matrix_country) * 2 # number of cells

round(table(matrix_country<5)/(nrow(matrix_country)*2),2) # 53% of cells have counts less than 5 

# running Fisher's exact test
fisher_test_country = fisher.test(matrix_country, simulate.p.value = TRUE)
fisher_test_country$p.value

# chi square test
chisq_country = chisq.test(data_prefs$sexual_orientation, data_prefs$country) 
chisq_country 
chisq_country$observed # observed counts
round(chisq_country$expected,2) # expected counts


# calculate Cramer's V
ci_cramersv(chisq_country, type = "chi-squared")
```

#### Language
```{r}
chisq_language = chisq.test(data_prefs$sexual_orientation, data_prefs$language) 
chisq_language
chisq_language$observed
round(chisq_language$expected,2)

# calculate Cramer's V
matrix_language = as.matrix(table(data_prefs$language, data_prefs$sexual_orientation), header=TRUE,
                     row.names=1)
round(cramerV(matrix_language, ci = T),2)
```

#### Age
```{r}
t.test(age ~ sexual_orientation, data = data_prefs,
       alternative = "two.sided",
       paired = TRUE)

# effect size
effsize::cohen.d(age ~ sexual_orientation, data = data_prefs, paired = TRUE)
```

#### Relationship status
```{r}
chisq_relstat = chisq.test(data_prefs$sexual_orientation, data_prefs$relationship_status) 
chisq_relstat
chisq_relstat$observed
round(chisq_relstat$expected,2)

# calculate Cramer's V
matrix_relstat = as.matrix(table(data_prefs$relationship_status, data_prefs$sexual_orientation), header=TRUE,
                     row.names=1)
round(cramerV(matrix_relstat, ci = T),2)
```

#### Relationship length
```{r}
t.test(relationship_length ~ sexual_orientation, data = data_prefs,
       alternative = "two.sided",
       paired = TRUE)

# effect size
effsize::cohen.d(relationship_length ~ sexual_orientation, data = data_prefs, paired = TRUE)
```

### For comparisons of matched pairs see markdown file 08_analyses

## 2. Asexuals vs. unmatched heterosexuals
### Covariates
#### Country
```{r}
# assumptions not met for chi square test: some cells contain less than 5 observations; therefore a Fisher's exact test is calculated
matrix_asex_v_umhetero_country = as.matrix(table(asex_v_umhetero_prefs$country, asex_v_umhetero_prefs$sexual_orientation), header=TRUE, row.names=1)
table(matrix_asex_v_umhetero_country < 5) # absolute number of cells containing less than 5 counts
nrow(matrix_asex_v_umhetero_country) * 2 # number of cells

round(table(matrix_asex_v_umhetero_country<5)/(nrow(matrix_asex_v_umhetero_country)*2),2) # 64% of cells have counts less than 5 

# running Fisher's exact test
fisher_test_asex_v_umhetero_country = fisher.test(matrix_asex_v_umhetero_country, simulate.p.value = TRUE)
fisher_test_asex_v_umhetero_country$p.value

# chi square test
chisq_asex_v_umhetero_country = chisq.test(asex_v_umhetero_prefs$sexual_orientation, asex_v_umhetero_prefs$country) 
chisq_asex_v_umhetero_country
chisq_asex_v_umhetero_country$observed # observed counts
round(chisq_asex_v_umhetero_country$expected,2) # expected counts


# calculate Cramer's V
ci_cramersv(chisq_asex_v_umhetero_country, type = "chi-squared")
```

#### Language
```{r}
language_asex_v_umhetero = chisq.test(asex_v_umhetero_prefs$language, asex_v_umhetero_prefs$sexual_orientation)
language_asex_v_umhetero
language_asex_v_umhetero$observed
round(language_asex_v_umhetero$expected,2)

# calculate Cramer's V
matrix_asex_v_umhetero_language = as.matrix(table(asex_v_umhetero_prefs$language, asex_v_umhetero_prefs$sexual_orientation), header=TRUE,
                     row.names=1)
round(cramerV(matrix_asex_v_umhetero_language, ci = T),2)
```

#### Age
```{r}
t.test(age ~ sexual_orientation, data = asex_v_umhetero_prefs,
       alternative = "two.sided",
       paired = FALSE)

# effect size
effsize::cohen.d(age ~ sexual_orientation, data = asex_v_umhetero_prefs, paired = FALSE)
```

#### Relationship status
```{r}
relstat_asex_v_umhetero = chisq.test(asex_v_umhetero_prefs$relationship_status, asex_v_umhetero_prefs$sexual_orientation)
relstat_asex_v_umhetero
relstat_asex_v_umhetero$observed
round(relstat_asex_v_umhetero$expected,2)

# calculate Cramer's V
matrix_asex_v_umhetero_relstat = as.matrix(table(asex_v_umhetero_prefs$relationship_status, asex_v_umhetero_prefs$sexual_orientation), header=TRUE,
                     row.names=1)
round(cramerV(matrix_asex_v_umhetero_relstat, ci = T),2)
```

#### Relationship length
```{r}
t.test(relationship_length ~ sexual_orientation, data = asex_v_umhetero_prefs,
       alternative = "two.sided",
       paired = FALSE)

# effect size
effsize::cohen.d(relationship_length ~ sexual_orientation, data = asex_v_umhetero_prefs, paired = FALSE)
```

### Outcomes
#### Preferred relationship options
```{r}
# quick overview
describeBy(asex_v_umhetero_rel, asex_v_umhetero_rel$sexual_orientation)
```

##### Interest non-sexual relationships
```{r}
t.test(interest_nonsexrel ~ sexual_orientation, 
       data = asex_v_umhetero_rel,
       alternative = "two.sided",
       paired = FALSE)

# effect size
effsize::cohen.d(interest_nonsexrel ~ sexual_orientation, data = asex_v_umhetero_rel , paired = F)
```

##### Interest sexual relationships
```{r}
t.test(interest_hookups ~ sexual_orientation, 
       data = asex_v_umhetero_rel,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(interest_hookups ~ sexual_orientation, data = data_rel, paired = F)
```

##### Interest non-monogamous relationships
```{r}
t.test(interest_nonmonrel ~ sexual_orientation, 
       data = asex_v_umhetero_rel,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(interest_nonmonrel ~ sexual_orientation, data = asex_v_umhetero_rel, paired = F)
```

##### Interest alternativ committed relationships
```{r}
t.test(interest_altrel ~ sexual_orientation, 
       data = asex_v_umhetero_rel,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(interest_altrel ~ sexual_orientation, data = asex_v_umhetero_rel, paired = F)
```

##### Interest single
```{r}
t.test(interest_single ~ sexual_orientation, 
       data = asex_v_umhetero_rel,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(interest_single ~ sexual_orientation, data = asex_v_umhetero_rel, paired = F)
```

##### Interest monogamous relationships
```{r}
t.test(interest_monrel ~ sexual_orientation, 
       data = asex_v_umhetero_rel,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(interest_monrel ~ sexual_orientation, data = asex_v_umhetero_rel, paired = F)
```

##### Interest parent 
```{r}
t.test(interest_parent ~ sexual_orientation, 
       data = asex_v_umhetero_rel,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(interest_parent ~ sexual_orientation, data = asex_v_umhetero_rel, paired = F)
```


#### Ideal partner preferences
```{r}
# quick overview
describeBy(asex_v_umhetero_prefs, asex_v_umhetero_prefs$sexual_orientation)
```

##### Confident-assertive
```{r}
t.test(pref_imp_ca ~ sexual_orientation, 
       data = asex_v_umhetero_prefs,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(pref_imp_ca ~ sexual_orientation, data = asex_v_umhetero_prefs, paired = FALSE)
```

##### Attractive
```{r}
t.test(pref_imp_att ~ sexual_orientation, 
       data = asex_v_umhetero_prefs,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(pref_imp_att ~ sexual_orientation, data = asex_v_umhetero_prefs, paired = FALSE)
```

##### Sexually experienced
```{r}
t.test(pref_imp_sexually_experienced ~ sexual_orientation, 
       data = asex_v_umhetero_prefs,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(pref_imp_sexually_experienced ~ sexual_orientation, data = asex_v_umhetero_prefs, paired = FALSE)
```

##### Kind-supportive
```{r}
t.test(pref_imp_ks ~ sexual_orientation, 
       data = asex_v_umhetero_prefs,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(pref_imp_ks ~ sexual_orientation, data = asex_v_umhetero_prefs, paired = FALSE)
```


##### Financially secure-successful
```{r}
t.test(pref_imp_fs ~ sexual_orientation, 
       data = asex_v_umhetero_prefs,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(pref_imp_fs ~ sexual_orientation, data = asex_v_umhetero_prefs, paired = FALSE)
```


##### Educated-intelligent
```{r}
t.test(pref_imp_ei ~ sexual_orientation, 
       data = asex_v_umhetero_prefs,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(pref_imp_ei ~ sexual_orientation, data = asex_v_umhetero_prefs, paired = FALSE)
```

#### Self-ratings
```{r}
# quick overview
describeBy(asex_v_umhetero_self, asex_v_umhetero_self$sexual_orientation)
```

##### Confident-assertive
```{r}
t.test(self_ca ~ sexual_orientation, 
       data = asex_v_umhetero_self,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(self_ca ~ sexual_orientation, data = asex_v_umhetero_self, paired = FALSE)
```

##### Attractive
```{r}
t.test(self_att ~ sexual_orientation, 
       data = asex_v_umhetero_self,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(self_att ~ sexual_orientation, data = asex_v_umhetero_self, paired = FALSE)
```

##### Sexually experienced
```{r}
t.test(self_sexually_experienced ~ sexual_orientation, 
       data = asex_v_umhetero_self,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(self_sexually_experienced ~ sexual_orientation, data = asex_v_umhetero_self, paired = FALSE)
```

##### Kind-supportive
```{r}
t.test(self_ks ~ sexual_orientation, 
       data = asex_v_umhetero_self,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(self_ks ~ sexual_orientation, data = asex_v_umhetero_self, paired = FALSE)
```


##### Financially secure-successful
```{r}
t.test(self_fs ~ sexual_orientation, 
       data = asex_v_umhetero_self,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(self_fs ~ sexual_orientation, data = asex_v_umhetero_self, paired = FALSE)
```


##### Educated-intelligent
```{r}
t.test(self_ei ~ sexual_orientation, 
       data = asex_v_umhetero_self,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(self_ei ~ sexual_orientation, data = asex_v_umhetero_self, paired = FALSE)
```


## 3. Matched heterosexuals v. unmatched heterosexuals
### Covariates
#### Country
```{r}
# assumptions not met for chi square test: some cells contain less than 5 observations; therefore a Fisher's exact test is calculated
matrix_hetero_v_umhetero_country = as.matrix(table(hetero_v_umhetero_prefs$country, hetero_v_umhetero_prefs$matched), header=TRUE, row.names=1)
table(matrix_hetero_v_umhetero_country < 5) # absolute number of cells containing less than 5 counts
nrow(matrix_hetero_v_umhetero_country) * 2 # number of cells

round(table(matrix_hetero_v_umhetero_country<5)/(nrow(matrix_hetero_v_umhetero_country)*2),2) # 64 % of cells have counts less than 5 

# running Fisher's exact test
fisher.test(matrix_hetero_v_umhetero_country, simulate.p.value = TRUE)

# chi square test
chisq_hetero_v_umhetero_country = chisq.test(matrix_hetero_v_umhetero_country)
chisq_hetero_v_umhetero_country
chisq_hetero_v_umhetero_country$observed # observed counts
round(chisq_hetero_v_umhetero_country$expected,2) # expected counts


# calculate Cramer's V
ci_cramersv(chisq_hetero_v_umhetero_country, type = "chi-squared")
```


#### Language
```{r}
language_hetero_v_umhetero = chisq.test(hetero_v_umhetero_prefs$language, hetero_v_umhetero_prefs$matched)
language_hetero_v_umhetero
language_hetero_v_umhetero$observed
round(language_hetero_v_umhetero$expected,2)

# calculate Cramer's V
matrix_hetero_v_umhetero_language = as.matrix(table(hetero_v_umhetero_prefs$language, hetero_v_umhetero_prefs$matched),
                                              header=TRUE, row.names=1)
round(cramerV(matrix_hetero_v_umhetero_language, ci = T),2)
```

#### Age
```{r}
t.test(age ~ matched, data = hetero_v_umhetero_prefs,
       alternative = "two.sided",
       paired = FALSE)

# effect size
effsize::cohen.d(age ~ matched, data = hetero_v_umhetero_prefs, paired = FALSE)
```

#### Relationship status
```{r}
relstat_hetero_v_umhetero = chisq.test(hetero_v_umhetero_prefs$relationship_status, hetero_v_umhetero_prefs$matched)
relstat_hetero_v_umhetero
relstat_hetero_v_umhetero$observed
round(relstat_hetero_v_umhetero$expected,2)

# calculate Cramer's V
matrix_hetero_v_umhetero_relstat = as.matrix(table(hetero_v_umhetero_prefs$relationship_status, hetero_v_umhetero_prefs$matched), header=TRUE,
                     row.names=1)
round(cramerV(matrix_hetero_v_umhetero_relstat, ci = T),2)
```

#### Relationship length
```{r}
t.test(relationship_length ~ matched, data = hetero_v_umhetero_prefs,
       alternative = "two.sided",
       paired = FALSE)

# effect size
effsize::cohen.d(relationship_length ~ matched, data = hetero_v_umhetero_prefs, paired = FALSE)
```

### Outcomes
#### Preferred relationship options
```{r}
# quick overview
describe(hetero_v_umhetero_rel)
```

##### Interest non-sexual relationships
```{r}
t.test(interest_nonsexrel ~ matched, 
       data = hetero_v_umhetero_rel,
       alternative = "two.sided",
       paired = FALSE)

# effect size
effsize::cohen.d(interest_nonsexrel ~ matched, data = hetero_v_umhetero_rel , paired = F)
```

##### Interest sexual relationships
```{r}
t.test(interest_hookups ~ matched, 
       data = hetero_v_umhetero_rel,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(interest_hookups ~ sexual_orientation, data = data_rel, paired = F)
```

##### Interest non-monogamous relationships
```{r}
t.test(interest_nonmonrel ~ matched, 
       data = hetero_v_umhetero_rel,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(interest_nonmonrel ~ matched, data = hetero_v_umhetero_rel, paired = F)
```

##### Interest alternativ committed relationships
```{r}
t.test(interest_altrel ~ matched, 
       data = hetero_v_umhetero_rel,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(interest_altrel ~ matched, data = hetero_v_umhetero_rel, paired = F)
```

##### Interest single
```{r}
t.test(interest_single ~ matched, 
       data = hetero_v_umhetero_rel,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(interest_single ~ matched, data = hetero_v_umhetero_rel, paired = F)
```

##### Interest monogamous relationships
```{r}
t.test(interest_monrel ~ matched, 
       data = hetero_v_umhetero_rel,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(interest_monrel ~ matched, data = hetero_v_umhetero_rel, paired = F)
```

##### Interest parent 
```{r}
t.test(interest_parent ~ matched, 
       data = hetero_v_umhetero_rel,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(interest_parent ~ matched, data = hetero_v_umhetero_rel, paired = F)
```


#### Ideal partner preferences
```{r}
# quick overview
describeBy(hetero_v_umhetero_prefs, hetero_v_umhetero_prefs$matched) 
```

##### Confident-assertive
```{r}
t.test(pref_imp_ca ~ matched, 
       data = hetero_v_umhetero_prefs,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(pref_imp_ca ~ matched, data = hetero_v_umhetero_prefs, paired = FALSE)
```

##### Attractive
```{r}
t.test(pref_imp_att ~ matched, 
       data = hetero_v_umhetero_prefs,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(pref_imp_att ~ matched, data = hetero_v_umhetero_prefs, paired = FALSE)
```

##### Sexually experienced
```{r}
t.test(pref_imp_sexually_experienced ~ matched, 
       data = hetero_v_umhetero_prefs,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(pref_imp_sexually_experienced ~ matched, data = hetero_v_umhetero_prefs, paired = FALSE)
```

##### Kind-supportive
```{r}
t.test(pref_imp_ks ~ matched, 
       data = hetero_v_umhetero_prefs,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(pref_imp_ks ~ matched, data = hetero_v_umhetero_prefs, paired = FALSE)
```


##### Financially secure-successful
```{r}
t.test(pref_imp_fs ~ matched, 
       data = hetero_v_umhetero_prefs,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(pref_imp_fs ~ matched, data = hetero_v_umhetero_prefs, paired = FALSE)
```


##### Educated-intelligent
```{r}
t.test(pref_imp_ei ~ matched, 
       data = hetero_v_umhetero_prefs,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(pref_imp_ei ~ matched, data = hetero_v_umhetero_prefs, paired = FALSE)
```

#### Self-ratings
```{r}
# quick overview
describeBy(hetero_v_umhetero_self, hetero_v_umhetero_self$matched) 
```

##### Confident-assertive
```{r}
t.test(self_ca ~ matched, 
       data = hetero_v_umhetero_self,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(self_ca ~ matched, data = hetero_v_umhetero_self, paired = FALSE)
```

##### Attractive
```{r}
t.test(self_att ~ matched, 
       data = hetero_v_umhetero_self,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(self_att ~ matched, data = hetero_v_umhetero_self, paired = FALSE)
```

##### Sexually experienced
```{r}
t.test(self_sexually_experienced ~ matched, 
       data = hetero_v_umhetero_self,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(self_sexually_experienced ~ matched, data = hetero_v_umhetero_self, paired = FALSE)
```

##### Kind-supportive
```{r}
t.test(self_ks ~ matched, 
       data = hetero_v_umhetero_self,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(self_ks ~ matched, data = hetero_v_umhetero_self, paired = FALSE)
```


##### Financially secure-successful
```{r}
t.test(self_fs ~ matched, 
       data = hetero_v_umhetero_self,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(self_fs ~ matched, data = hetero_v_umhetero_self, paired = FALSE)
```


##### Educated-intelligent
```{r}
t.test(self_ei ~ matched, 
       data = hetero_v_umhetero_self,
       alternative = "two.sided",
       paired = FALSE)


# effect size
effsize::cohen.d(self_ei ~ matched, data = hetero_v_umhetero_self, paired = FALSE)
```
