==> 12_analyses_additional.Rmd <==
---
title: "Additional Analyses"
output: 
  html_document:
    toc: true
    toc_depth: 6
    toc_float: true
    code_folding: 'show'
editor_options: 
  chunk_output_type: console
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(warnings = FALSE)
```

## Library
```{r}
library(dplyr)
library(ggplot2)
library(MatchIt)
library(cobalt)
library(psych)
library(effsize)
library(forestplot)

## set options for bal.tab function in cobalt package
set.cobalt.options(binary = "std") # for binary covariates used in the matching procedure the bal.tab function will now display standardized mean differences instead of the raw difference
```


## Load data
```{r}
data_main = read.csv(file = "data/data_main.csv")[,-1]
```

## 1. PSM for partner preferences
Check for missings in covariates to be matched
- psm cannot handle missings
- covariates: language, age, country, relationship status, relationship length
- additionally match for: interest_nonsexrel, interest_monrel, interest_single
```{r}
# MatchIt does not allow missing values
data_main_nomiss_add <- data_main %>% select(id, language, age, sex, country, relationship_status, relationship_length, sexual_orientation,
                                         pref_imp_ks, pref_imp_att, pref_imp_fs, pref_imp_ca,
                                         pref_imp_ei, pref_imp_sexually_experienced,
                                         interest_nonsexrel, interest_monrel, interest_single) %>%
                                  na.omit()

table(is.na(data_main_nomiss_add)) # no more NAs in data set
```

Check if correct variable type 
```{r}
typeof(data_main_nomiss_add$sexual_orientation) # is variable of type character, but for PSM it needs to be a factor
data_main_nomiss_add$sexual_orientation = as.factor(data_main_nomiss_add$sexual_orientation) # convert to factor
levels(data_main_nomiss_add$sexual_orientation) # check levels of factor
data_main_nomiss_add$sexual_orientation = factor(data_main_nomiss_add$sexual_orientation,
                                             levels=rev(levels(data_main_nomiss_add$sexual_orientation))) # reverse order of factor levels 
levels(data_main_nomiss_add$sexual_orientation) # check factor level order
```


### Matching with NNM
```{r}
nnm_add <- matchit(sexual_orientation ~ language + age + country + relationship_status + 
                     relationship_length + interest_nonsexrel + interest_monrel +
                     interest_single, 
                   method = "nearest", data = data_main_nomiss_add)

nnm_add_data <- match.data(nnm_add)

nrow(nnm_add_data) # 636 observations
table(nnm_add_data$subclass) # data includes only pairs of two observations
dim(nnm_add_data) # 636 observations and 20 variables
```

### Check balance
```{r}
summary_nnm_add = summary(nnm_add, standardize = T)
print(summary_nnm_add)

bal.tab(nnm_add, m.threshold = 0.1, un = T, binary = "std") #smd
bal.tab(nnm_add, v.threshold = 2, un = T) #vr (can only be calculated for continuous variables)
```
No asexual unmatched and all variables balanced.

### Visualize
```{r}
# Q-Q plot
plot(nnm_add, type = "qq", interactive = FALSE, which.xs = c("age", "relationship_length", "interest_nonsexrel", "interest_monrel", "interest_single"))
# looks better after matching

# Jitter plot
plot(nnm_add, type = "jitter", interactive = FALSE)
# one observation in treated sample has a higher propensity score than any of the control units; however, most observations seem to fall within the area of common support

# distributional density/histogram
bal.plot(nnm_add, var.name = "age", which = "both", grid = T)
bal.plot(nnm_add, var.name = "relationship_length", which = "both", grid = T)
bal.plot(nnm_add, var.name = "interest_nonsexrel", which = "both", grid = T)
bal.plot(nnm_add, var.name = "interest_monrel", which = "both", grid = T)
bal.plot(nnm_add, var.name = "interest_single", which = "both", grid = T)

bal.plot(nnm_add, var.name = "relationship_status", which = "both", grid = T)
bal.plot(nnm_add, var.name = "language", which = "both", grid = T)
bal.plot(nnm_add, var.name = "country", which = "both", grid = T) # not readable

# remove country variable for love plot
nnm_add_without_country = nnm_add
nnm_add_without_country$X[c("country")] <- NULL 

love.plot(bal.tab(nnm_add_without_country, m.threshold = 0.1),
          stat = "mean.diffs",
          grid = T,
          stars = "raw",
          abs = F,
          binary = "std") 


## love plot using alternate variable names
# Using alternate variable names
v <- data.frame(old = c("distance", "language_chinese", "language_japanese", "language_german","language_english",
                        "language_italian", "language_danish", "language_spanish", "language_russian",
                        "language_portuguese", "language_french", "age", "relationship_status_Long-term", 
                        "relationship_status_No relationship", "relationship_status_Short-term",
                        "relationship_length", "interest_nonsexrel", "interest_monrel", "interest_single"),
                new = c("Propensity scores", "Chinese", "Japanese", "German", "English", "Italian", "Danish", 
                        "Spanish", "Russian", "Portuguese", "French", "Long-term relationship", "No relationship", 
                        "Dating/Sexual Relationship","Age",  "Relationship length", "Interest non-sexual, 
                        romantic relationships", "Interest monogamous relationships", "Interest single"))

love.plot(bal.tab(nnm_add_without_country, m.threshold = 0.1, binary = "std"),
          stat = "mean.diffs",
          grid = T,
          stars = "raw",
          abs = F,
          sample.names = c("Before Matching", "After Matching"),
          var.names = v) 
```


### Recalculate t-test for differences in partner preferences based on this psm sample controlling for having no interest in any long-term relationship
Sort data by subclass number within sexual_orientation (so that all asexual participants appear first sorted by subclass)
```{r}
data_add = nnm_add_data %>% 
        group_by(sexual_orientation) %>%
        arrange(subclass, .by_group = TRUE) 

typeof(data_add$sexual_orientation) # sexual_orientation is a factor
levels(data_add$sexual_orientation) # check factor level order
# for analyses, I reverse the factor levels so that 0 = Asexuals and 1 = Straight/Heterosexual 
data_add$sexual_orientation =  factor(data_add$sexual_orientation,
                                             levels=rev(levels(data_add$sexual_orientation))) # reverse order of factor levels 
levels(data_add$sexual_orientation) # check factor level order
```


#### Quick overview
```{r}
# mean
data_add %>% group_by(sexual_orientation) %>% 
  summarise_at(c("pref_imp_ca", 
                 "pref_imp_att",
                 "pref_imp_sexually_experienced",
                 "pref_imp_ks",
                 "pref_imp_fs", 
                 "pref_imp_ei"), mean, na.rm = T)
# sd
data_add %>% group_by(sexual_orientation) %>% 
  summarise_at(c("pref_imp_ca", 
                 "pref_imp_att",
                 "pref_imp_sexually_experienced",
                 "pref_imp_ks",
                 "pref_imp_fs", 
                 "pref_imp_ei"), sd, na.rm = T)

# Visualize
# confident-assertive
ggplot(data_add, aes(x = pref_imp_ca, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge")


# attractive
ggplot(data_add, aes(x = pref_imp_att, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge")

# sexually experienced
ggplot(data_add, aes(x = pref_imp_sexually_experienced, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge")

# kind-supportive
ggplot(data_add, aes(x = pref_imp_ks, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge")

# financially secure-successful
ggplot(data_add, aes(x = pref_imp_fs, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge")

# educated-intelligent
ggplot(data_add, aes(x = pref_imp_ei, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge")
```

#### Confident-assertive
```{r}
t.test(pref_imp_ca ~ sexual_orientation, 
       data = data_add,
       alternative = "less",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_ca ~ sexual_orientation, data = data_add, paired = T)
```

#### Attractive
```{r}
t.test(pref_imp_att ~ sexual_orientation, 
       data = data_add,
       alternative = "less",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_att ~ sexual_orientation, data = data_add, paired = T)
```

#### Sexually experienced
```{r}
t.test(pref_imp_sexually_experienced ~ sexual_orientation, 
       data = data_add,
       alternative = "less",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_sexually_experienced ~ sexual_orientation, data = data_add, paired = T)
```

#### Exploratory analyses
##### Kind-supportive
```{r}
t.test(pref_imp_ks ~ sexual_orientation, 
       data = data_add,
       alternative = "two.sided",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_ks ~ sexual_orientation, data = data_add, paired = T)
```


##### Financially secure-successful
```{r}
t.test(pref_imp_fs ~ sexual_orientation, 
       data = data_add,
       alternative = "two.sided",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_fs ~ sexual_orientation, data = data_add, paired = T)
```


##### Educated-intelligent
```{r}
t.test(pref_imp_ei ~ sexual_orientation, 
       data = data_add,
       alternative = "two.sided",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_ei ~ sexual_orientation, data = data_add, paired = T)
```


### Create forest plot
```{r, fig.width = 10, fig.height = 6}
base_data_add <- tibble(mean  = c(-0.45, -0.49, -0.61, -0.14, -0.50, 0.03), 
                    lower = c(-0.61, -0.66, -0.77, -0.30, -0.66, -0.12),
                    upper = c(-0.30, -0.32, -0.46, 0.02, -0.33, 0.19),
                    Partner_preferences = c("Confident-assertive", "Attractive", 
                                            "Sexually experienced", "Kind-supportive", 
                                            "Financially secure-successful", "Educated-intelligent"),
                    d = c("-0.45", "-0.49", "-0.61", "-0.14", "-0.50", "0.03"))

header_add <- tibble(Partner_preferences = c("", "Partner preferences"),
                d = c("", "d"),
                 summary = TRUE)

partner_preferences_add <- bind_rows(header_add,
                                base_data_add)



partner_preferences_add %>% 
  forestplot(labeltext = c(Partner_preferences, d), 
             graph.pos = 3,
              clip = c(-1,2),
             lty.ci = 2,
             boxsize = .25,
              is.summary = summary,
              grid = structure(c(-.26, .26), 
                              gp = gpar(lty = 2, col = "darkblue")), 
             xticks = c(-1.5, -.26, -0.5,0, .26, .5, 1, 1.5),
             xlab = "Effect size (Cohen's d)",
             txt_gp = fpTxtGp(ticks=gpar(cex=.75), xlab=gpar(cex=1)),
             col = fpColors(box = "royalblue",
                            line = "black",
                            summary = "royalblue"),
             title = "Controlling for disinterest in long-term relationships (N = 636)",
             graphwidth = unit(150, "mm"))
```


## 2.  PSM including interest in purely sexual relationships
Check for missings in covariates to be matched
- psm cannot handle missings
- covariates: language, age, country, relationship status, relationship length
- additionally match for: interest_nonsexrel, interest_monrel, interest_single & interest_hookups
```{r}
# MatchIt does not allow missing values
data_main_nomiss_add2 <- data_main %>% select(id, language, age, sex, country, relationship_status, 
                                              relationship_length, sexual_orientation,
                                              pref_imp_ks, pref_imp_att, pref_imp_fs, pref_imp_ca,
                                              pref_imp_ei, pref_imp_sexually_experienced,
                                              interest_nonsexrel, interest_monrel, interest_single,
                                              interest_hookups) %>%
                                  na.omit()

table(is.na(data_main_nomiss_add2)) # no more NAs in data set
```

Check if correct variable type 
```{r}
typeof(data_main_nomiss_add2$sexual_orientation) # is variable of type character, but for PSM it needs to be a factor
data_main_nomiss_add2$sexual_orientation = as.factor(data_main_nomiss_add2$sexual_orientation) # convert to factor
levels(data_main_nomiss_add2$sexual_orientation) # check levels of factor
data_main_nomiss_add2$sexual_orientation = factor(data_main_nomiss_add2$sexual_orientation,
                                             levels=rev(levels(data_main_nomiss_add2$sexual_orientation))) # reverse order of factor levels 
levels(data_main_nomiss_add2$sexual_orientation) # check factor level order
```

### Matching with NNM
```{r}
nnm_add2 <- matchit(sexual_orientation ~ language + age + country + relationship_status + 
                      relationship_length + interest_nonsexrel + interest_monrel + 
                      interest_single + interest_hookups, 
                    method = "nearest", data = data_main_nomiss_add2)

nnm_add2_data <- match.data(nnm_add2)

nrow(nnm_add2_data) # 636 observations
table(nnm_add2_data$subclass) # data includes only pairs of two observations
dim(nnm_add2_data) # 636 observations and 21 variables
```

### Check balance
```{r}
summary_nnm_add2 = summary(nnm_add2, standardize = T)
print(summary_nnm_add2)
tail(summary_nnm_add2$sum.matched) #look at balance properties for interest variables

bal.tab(nnm_add2, m.threshold = 0.1, un = T, binary = "std") #smd
bal.tab(nnm_add2, v.threshold = 2, un = T) #vr (can only be calculated for continuous variables)
```
No asexual unmatched. Based on smds 5 covariates not balanced (country_Singapore, country_Japan, country_France, country_Argentina, language_japanese). Based on vr, all continuous covariates balanced. 

### Visualize
```{r}
# Q-Q plot
plot(nnm_add2, type = "qq", interactive = FALSE, which.xs = c("age", "relationship_length", "interest_nonsexrel", "interest_monrel", "interest_single"))
# looks better after matching

# Jitter plot
plot(nnm_add2, type = "jitter", interactive = FALSE)
# two observations in treated sample has a higher propensity score than any of the control units; however, most observations seem to fall within the area of common support

# distributional density/histogram
bal.plot(nnm_add2, var.name = "age", which = "both", grid = T)
bal.plot(nnm_add2, var.name = "relationship_length", which = "both", grid = T)
bal.plot(nnm_add2, var.name = "interest_nonsexrel", which = "both", grid = T)
bal.plot(nnm_add2, var.name = "interest_monrel", which = "both", grid = T)
bal.plot(nnm_add2, var.name = "interest_single", which = "both", grid = T)
bal.plot(nnm_add2, var.name = "interest_hookups", which = "both", grid = T)

bal.plot(nnm_add2, var.name = "relationship_status", which = "both", grid = T)
bal.plot(nnm_add2, var.name = "language", which = "both", grid = T)
bal.plot(nnm_add2, var.name = "country", which = "both", grid = T) # not readable

# remove country variable for love plot
nnm_add2_without_country = nnm_add2
nnm_add2_without_country$X[c("country")] <- NULL 

love.plot(bal.tab(nnm_add2_without_country, m.threshold = 0.1),
          stat = "mean.diffs",
          grid = T,
          stars = "raw",
          abs = F,
          binary = "std") 


## love plot using alternate variable names
# Using alternate variable names
v <- data.frame(old = c("distance", "language_chinese", "language_japanese", "language_german","language_english",
                        "language_italian", "language_danish", "language_spanish", "language_russian", 
                        "language_portuguese", "language_french", "age", "relationship_status_Long-term", 
                        "relationship_status_No relationship", "relationship_status_Short-term",
                        "relationship_length", "interest_nonsexrel", "interest_monrel", "interest_single",
                        "interest_hookups"),
                new = c("Propensity scores", "Chinese", "Japanese", "German", "English", "Italian", "Danish", 
                        "Spanish", "Russian", "Portuguese", "French", "Long-term relationship", "No relationship",
                        "Dating/Sexual Relationship","Age",  "Relationship length", "Interest non-sexual,
                        romantic relationships", "Interest monogamous relationships", "Interest single", 
                        "Interest sexual, non-romantic relationships"))

love.plot(bal.tab(nnm_add2_without_country, m.threshold = 0.1, binary = "std"),
          stat = "mean.diffs",
          grid = T,
          stars = "raw",
          abs = F,
          sample.names = c("Before Matching", "After Matching"),
          var.names = v) 
```

### Recalculate t-test for differences in partner preferences based on this psm sample controlling for having no interest in any long-term relationship
Sort data by subclass number within sexual_orientation
```{r}
data_add2 = nnm_add2_data %>% 
        group_by(sexual_orientation) %>%
        arrange(subclass, .by_group = TRUE)

levels(data_add2$sexual_orientation) # check factor level
# for analyses, I reverse the factor levels so that 0 = Asexuals and 1 = Straight/Heterosexual 
data_add2$sexual_orientation =  factor(data_add2$sexual_orientation,
                                             levels=rev(levels(data_add2$sexual_orientation))) # reverse order of factor levels 
levels(data_add2$sexual_orientation) # check factor level order

```

#### Quick overview
```{r}
# mean
data_add2 %>% group_by(sexual_orientation) %>% 
  summarise_at(c("pref_imp_ca", 
                 "pref_imp_att",
                 "pref_imp_sexually_experienced",
                 "pref_imp_ks",
                 "pref_imp_fs", 
                 "pref_imp_ei"), mean, na.rm = T)
# sd
data_add2 %>% group_by(sexual_orientation) %>% 
  summarise_at(c("pref_imp_ca", 
                 "pref_imp_att",
                 "pref_imp_sexually_experienced",
                 "pref_imp_ks",
                 "pref_imp_fs", 
                 "pref_imp_ei"), sd, na.rm = T)

# Visualize
# confident-assertive
ggplot(data_add2, aes(x = pref_imp_ca, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge")


# attractive
ggplot(data_add2, aes(x = pref_imp_att, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge")

# sexually experienced
ggplot(data_add2, aes(x = pref_imp_sexually_experienced, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge")

# kind-supportive
ggplot(data_add2, aes(x = pref_imp_ks, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge")

# financially secure-successful
ggplot(data_add2, aes(x = pref_imp_fs, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge")

# educated-intelligent
ggplot(data_add2, aes(x = pref_imp_ei, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge")
```

#### Confident-assertive
```{r}
t.test(pref_imp_ca ~ sexual_orientation, 
       data = data_add2,
       alternative = "less",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_ca ~ sexual_orientation, data = data_add2, paired = T)
```

#### Attractive
```{r}
t.test(pref_imp_att ~ sexual_orientation, 
       data = data_add2,
       alternative = "less",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_att ~ sexual_orientation, data = data_add2, paired = T)
```

#### Sexually experienced
```{r}
t.test(pref_imp_sexually_experienced ~ sexual_orientation, 
       data = data_add2,
       alternative = "less",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_sexually_experienced ~ sexual_orientation, data = data_add2, paired = T)
```

#### Exploratory analyses
##### Kind-supportive
```{r}
t.test(pref_imp_ks ~ sexual_orientation, 
       data = data_add2,
       alternative = "two.sided",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_ks ~ sexual_orientation, data = data_add2, paired = T)
```


##### Financially secure-successful
```{r}
t.test(pref_imp_fs ~ sexual_orientation, 
       data = data_add2,
       alternative = "two.sided",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_fs ~ sexual_orientation, data = data_add2, paired = T)
```


##### Educated-intelligent
```{r}
t.test(pref_imp_ei ~ sexual_orientation, 
       data = data_add2,
       alternative = "two.sided",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_ei ~ sexual_orientation, data = data_add2, paired = T)
```

### Create forest plot
```{r, fig.width = 10, fig.height = 6}
base_data_add2 <- tibble(mean  = c(-0.49, -0.43, -0.37, -0.23, -0.44, -0.04), 
                    lower = c(-0.65, -0.60, -0.52, -0.38, -0.60, -0.19),
                    upper = c(-0.33, -0.27, -0.22, -0.08, -0.28, 0.11),
                    Partner_preferences = c("Confident-assertive", "Attractive", 
                                            "Sexually experienced", "Kind-supportive",
                                            "Financially secure-successful", "Educated-intelligent"),
                    d = c("-0.49", "-0.43", "-0.37", "-0.23", "-0.44", "-0.04"))

header_add2 <- tibble(Partner_preferences = c("", "Partner preferences"),
                d = c("", "d"),
                 summary = TRUE)

partner_preferences_add2 <- bind_rows(header_add2,
                                base_data_add2)



partner_preferences_add2 %>% 
  forestplot(labeltext = c(Partner_preferences, d), 
             graph.pos = 3,
              clip = c(-1,2),
             lty.ci = 2,
             boxsize = .25,
              is.summary = summary,
              grid = structure(c(-.26, .26), 
                              gp = gpar(lty = 2, col = "darkblue")), 
             xticks = c(-1.5, -.26, -0.5,0, .26, .5, 1, 1.5),
             xlab = "Effect size (Cohen's d)",
             txt_gp = fpTxtGp(ticks=gpar(cex=.75), xlab=gpar(cex=1)),
             col = fpColors(box = "royalblue",
                            line = "black",
                            summary = "royalblue"),
             title = "+ Interest in purely sexual relationships (N = 636)",
             graphwidth = unit(150, "mm"))
```

## 3. PSM with parenting interest
Check for missings in covariates to be matched
- psm cannot handle missings
- covariates: language, age, country, relationship status, relationship length
- additionally match for: interest_nonsexrel, interest_monrel, interest_single & interest_hookups
```{r}
# MatchIt does not allow missing values
data_main_nomiss_add3 <- data_main %>% select(id, language, age, sex, country, relationship_status,
                                              relationship_length, sexual_orientation,
                                              pref_imp_ks, pref_imp_att, pref_imp_fs, pref_imp_ca,
                                              pref_imp_ei, pref_imp_sexually_experienced,
                                              interest_nonsexrel, interest_monrel, interest_single,
                                              interest_parent) %>%
                                  na.omit()

table(is.na(data_main_nomiss_add3)) # no more NAs in data set
```

Check if correct variable type 
```{r}
typeof(data_main_nomiss_add3$sexual_orientation) # is variable of type character, but for PSM it needs to be a factor
data_main_nomiss_add3$sexual_orientation = as.factor(data_main_nomiss_add3$sexual_orientation) # convert to factor
levels(data_main_nomiss_add3$sexual_orientation) # check levels of factor
data_main_nomiss_add3$sexual_orientation = factor(data_main_nomiss_add3$sexual_orientation,
                                             levels=rev(levels(data_main_nomiss_add3$sexual_orientation))) # reverse order of factor levels 
levels(data_main_nomiss_add3$sexual_orientation) # check factor level order
```


### Matching with NNM
```{r}
nnm_add3 <- matchit(sexual_orientation ~ language + age + country + relationship_status + 
                      relationship_length + interest_nonsexrel + interest_monrel + 
                      interest_single + interest_parent, 
                    method = "nearest", data = data_main_nomiss_add3)

nnm_add3_data <- match.data(nnm_add3)

nrow(nnm_add3_data) # 634 observations
table(nnm_add3_data$subclass) # data includes only pairs of two observations
dim(nnm_add3_data) # 634 observations and 21 variables
```

### Check balance
```{r}
summary_nnm_add3 = summary(nnm_add3, standardize = T)
print(summary_nnm_add3)
tail(summary_nnm_add3$sum.matched) #look at balance properties for interest variables

bal.tab(nnm_add3, m.threshold = 0.1, un = T, binary = "std") #smd
bal.tab(nnm_add3, v.threshold = 2, un = T) #vr (can only be calculated for continuous variables)
```
No asexual unmatched. Based on smds 2 covariates not balanced (country_Denmark and langugage_danish). Based on vr, relationship_length is not balanced (vr = .29)

### Visualize
```{r}
# Q-Q plot
plot(nnm_add3, type = "qq", interactive = FALSE, which.xs = c("age", "relationship_length", "interest_nonsexrel", "interest_monrel", "interest_single", "interest_parent"))
# looks better after matching

# Jitter plot
plot(nnm_add3, type = "jitter", interactive = FALSE)

# distributional density/histogram
bal.plot(nnm_add3, var.name = "age", which = "both", grid = T)
bal.plot(nnm_add3, var.name = "relationship_length", which = "both", grid = T)
bal.plot(nnm_add3, var.name = "interest_nonsexrel", which = "both", grid = T)
bal.plot(nnm_add3, var.name = "interest_monrel", which = "both", grid = T)
bal.plot(nnm_add3, var.name = "interest_single", which = "both", grid = T)
bal.plot(nnm_add3, var.name = "interest_parent", which = "both", grid = T)

bal.plot(nnm_add3, var.name = "relationship_status", which = "both", grid = T)
bal.plot(nnm_add3, var.name = "language", which = "both", grid = T)
bal.plot(nnm_add3, var.name = "country", which = "both", grid = T) # not readable

# remove country variable for love plot
nnm_add3_without_country = nnm_add3
nnm_add3_without_country$X[c("country")] <- NULL 

love.plot(bal.tab(nnm_add3_without_country, m.threshold = 0.1),
          stat = "mean.diffs",
          grid = T,
          stars = "raw",
          abs = F,
          binary = "std") 


## love plot using alternate variable names
# Using alternate variable names
v <- data.frame(old = c("distance", "language_chinese", "language_japanese", "language_german","language_english",
                        "language_italian", "language_danish", "language_spanish", "language_russian", 
                        "language_portuguese", "language_french", "age", "relationship_status_Long-term", 
                        "relationship_status_No relationship", "relationship_status_Short-term",
                        "relationship_length", "interest_nonsexrel", "interest_monrel", "interest_single", 
                        "interest_parent"),
                new = c("Propensity scores", "Chinese", "Japanese", "German", "English", "Italian", "Danish",
                        "Spanish", "Russian", "Portuguese", "French", "Long-term relationship", 
                        "No relationship", "Dating/Sexual Relationship","Age",  "Relationship length", 
                        "Interest non-sexual, romantic relationships", "Interest monogamous relationships", 
                        "Interest single", "Interest becoming a parent"))

love.plot(bal.tab(nnm_add3_without_country, m.threshold = 0.1, binary = "std"),
          stat = "mean.diffs",
          grid = T,
          stars = "raw",
          abs = F,
          sample.names = c("Before Matching", "After Matching"),
          var.names = v) 
```


### Recalculate t-test for differences in partner preferences based on this psm sample controlling for having no interest in any long-term relationship
Sort data by subclass number within sexual_orientation
```{r}
data_add3 = nnm_add3_data %>% 
        group_by(sexual_orientation) %>%
        arrange(subclass, .by_group = TRUE)

levels(data_add3$sexual_orientation) # check factor level
# for analyses, I reverse the factor levels so that 0 = Asexuals and 1 = Straight/Heterosexual 
data_add3$sexual_orientation =  factor(data_add3$sexual_orientation,
                                             levels=rev(levels(data_add3$sexual_orientation))) # reverse order of factor levels 
levels(data_add3$sexual_orientation) # check factor level order

```

#### Quick overview
```{r}
# mean
data_add3 %>% group_by(sexual_orientation) %>% 
  summarise_at(c("pref_imp_ca", 
                 "pref_imp_att",
                 "pref_imp_sexually_experienced",
                 "pref_imp_ks",
                 "pref_imp_fs", 
                 "pref_imp_ei"), mean, na.rm = T)
# sd
data_add3 %>% group_by(sexual_orientation) %>% 
  summarise_at(c("pref_imp_ca", 
                 "pref_imp_att",
                 "pref_imp_sexually_experienced",
                 "pref_imp_ks",
                 "pref_imp_fs", 
                 "pref_imp_ei"), sd, na.rm = T)

# Visualize
# confident-assertive
ggplot(data_add3, aes(x = pref_imp_ca, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge")


# attractive
ggplot(data_add3, aes(x = pref_imp_att, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge")

# sexually experienced
ggplot(data_add3, aes(x = pref_imp_sexually_experienced, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge")

# kind-supportive
ggplot(data_add3, aes(x = pref_imp_ks, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge")

# financially secure-successful
ggplot(data_add3, aes(x = pref_imp_fs, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge")

# educated-intelligent
ggplot(data_add3, aes(x = pref_imp_ei, fill = sexual_orientation)) + 
  geom_histogram(position = "dodge")
```

#### Confident-assertive
```{r}
t.test(pref_imp_ca ~ sexual_orientation, 
       data = data_add3,
       alternative = "less",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_ca ~ sexual_orientation, data = data_add3, paired = T)
```

#### Attractive
```{r}
t.test(pref_imp_att ~ sexual_orientation, 
       data = data_add3,
       alternative = "less",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_att ~ sexual_orientation, data = data_add3, paired = T)
```

#### Sexually experienced
```{r}
t.test(pref_imp_sexually_experienced ~ sexual_orientation, 
       data = data_add3,
       alternative = "less",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_sexually_experienced ~ sexual_orientation, data = data_add3, paired = T)
```

#### Exploratory analyses
##### Kind-supportive
```{r}
t.test(pref_imp_ks ~ sexual_orientation, 
       data = data_add3,
       alternative = "two.sided",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_ks ~ sexual_orientation, data = data_add3, paired = T)
```


##### Financially secure-successful
```{r}
t.test(pref_imp_fs ~ sexual_orientation, 
       data = data_add3,
       alternative = "two.sided",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_fs ~ sexual_orientation, data = data_add3, paired = T)
```


##### Educated-intelligent
```{r}
t.test(pref_imp_ei ~ sexual_orientation, 
       data = data_add3,
       alternative = "two.sided",
       paired = TRUE)


# effect size
effsize::cohen.d(pref_imp_ei ~ sexual_orientation, data = nnm_add3_data, paired = T)
```

### Create forest plot
```{r, fig.width = 10, fig.height = 6}
base_data_add3 <- tibble(mean  = c(-0.35, -0.43, -0.59, -0.06, -0.33, 0.05), 
                    lower = c(-0.51, -0.59, -0.76, -0.22, -0.49, -0.11),
                    upper = c(-0.20, -0.27, -0.43, 0.10, -0.17, 0.20),
                    Partner_preferences = c("Confident-assertive", "Attractive", 
                                            "Sexually experienced", "Kind-supportive", 
                                            "Financially secure-successful", "Educated-intelligent"),
                    d = c("-0.35", "-0.43", "-0.59", "-0.06", "-0.33", "0.05"))

header_add3 <- tibble(Partner_preferences = c("", "Partner preferences"),
                d = c("", "d"),
                 summary = TRUE)

partner_preferences_add3 <- bind_rows(header_add3,
                                base_data_add3)



partner_preferences_add3 %>% 
  forestplot(labeltext = c(Partner_preferences, d), 
             graph.pos = 3,
              clip = c(-1,2),
             lty.ci = 2,
             boxsize = .25,
              is.summary = summary,
              grid = structure(c(-.26, .26), 
                              gp = gpar(lty = 2, col = "darkblue")), 
             xticks = c(-1.5, -.26, -0.5,0, .26, .5, 1, 1.5),
             xlab = "Effect size (Cohen's d)",
             txt_gp = fpTxtGp(ticks=gpar(cex=.75), xlab=gpar(cex=1)),
             col = fpColors(box = "royalblue",
                            line = "black",
                            summary = "royalblue"),
             title = "+ Parenting interest (N = 634)",
             graphwidth = unit(150, "mm"))
```
