==> 13_lifetime_sexual_partners.Rmd <==
---
title: "Number of Lifetime Sexual Partners"
output: 
  html_document:
    toc: true
    toc_depth: 6
    toc_float: true
    code_folding: 'show'
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(warning = FALSE)
```
*[Disclaimer]*
Originally, the sexual partner variables were added post hoc to the data set used in all previous analyses. However, for the simulated data set, they are already included. Thus, if you want to replicate the code below, select "fulldata_selected" instead of "fulldata_selected_pus_sexual_partner_variable" to load data with the sexual partner variable. 

## Library
```{r}
library(ggplot2)
library(dplyr)
library(effsize)
```

## Load data
```{r}
# load data_prefs and data_self
data_rel = read.csv(file = "data/data_rel.csv")[,-1]
data_prefs = read.csv(file = "data/data_prefs.csv")[,-1]
data_self = read.csv(file = "data/data_self.csv")[,-1]

# load data with lifetime sexual partners variable
fulldata_selected_plus_sexual_partner_variable = read.csv(file = "data/fulldata_selected_plus_sexual_partner_variable.csv")[,-1]
names(fulldata_selected_plus_sexual_partner_variable)
sexual_partner = fulldata_selected_plus_sexual_partner_variable %>% select("total_sex_partners", "total_sex_partners2")
```


Add id variable and add lifetime sexual partner variable to data sets
```{r}
# add id variable to sexual partner data set
sexual_partner$id = 1:94738

# add variables to data
data_rel_plus_sexual_partner_variable = left_join(data_rel, sexual_partner)
data_prefs_plus_sexual_partner_variable = left_join(data_prefs, sexual_partner) # joined by id variable
data_self_plus_sexual_partner_variable = left_join(data_self, sexual_partner) # joined by id variable
```


## Partner preferences: look at sexual partners variables
```{r}
# mean
data_prefs_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), mean, na.rm = T)

# sd
data_prefs_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), sd, na.rm = T)

# min
data_prefs_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), min, na.rm = T)

# max
data_prefs_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), max, na.rm = T)

# range
data_prefs_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), range, na.rm = T)

# median
data_prefs_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), median, na.rm = T)

# counts
table(data_prefs_plus_sexual_partner_variable$total_sex_partners)
table(data_prefs_plus_sexual_partner_variable$total_sex_partners2)

# correlation
cor.test(data_prefs_plus_sexual_partner_variable$total_sex_partners, data_prefs_plus_sexual_partner_variable$total_sex_partners2, method = "pearson", na.rm = T)
```
Both variables are highly correlated (r = 0.99) - as expected. total_sex_partners_2 will be used for the following analysis.


### Correlation between number of total sex partners and preference for a sexually experienced partner
```{r}
cor.test(data_prefs_plus_sexual_partner_variable$total_sex_partners2, data_prefs_plus_sexual_partner_variable$pref_imp_sexually_experienced, method = "pearson", na.rm = T)
```
r = 0.26

### t-test difference between asexuals and heterosexuals in total number of sex partners
```{r}
t.test(total_sex_partners2 ~ sexual_orientation, data = data_prefs_plus_sexual_partner_variable,
       alternative = "two.sided",
       paired = F)

effsize::cohen.d(total_sex_partners2 ~ sexual_orientation, data = data_prefs_plus_sexual_partner_variable,
       paired = F)
```


## Self-ratings: look at sexual partners variables
```{r}
# mean
data_self_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), mean, na.rm = T)

# sd
data_self_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), sd, na.rm = T)

# min
data_self_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), min, na.rm = T)

# max
data_self_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), max, na.rm = T)

# range
data_self_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), range, na.rm = T)

# median
data_self_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), median, na.rm = T)

# counts
table(data_self_plus_sexual_partner_variable$total_sex_partners)
table(data_self_plus_sexual_partner_variable$total_sex_partners2)

# correlation
cor.test(data_self_plus_sexual_partner_variable$total_sex_partners, data_self_plus_sexual_partner_variable$total_sex_partners2, method = "pearson", na.rm = T)
```
Both variables are highly correlated (r = 0.99).

### Correlation between total number of sex partners and being sexually experienced 
```{r}
cor.test(data_self_plus_sexual_partner_variable$total_sex_partners2, data_self_plus_sexual_partner_variable$self_sexually_experienced, method = "pearson", na.rm = T)
```

### t-test difference between asexuals and heterosexuals in total number of sex partners
```{r}
t.test(total_sex_partners2 ~ sexual_orientation, data = data_self_plus_sexual_partner_variable,
       alternative = "two.sided",
       paired = F)

effsize::cohen.d(total_sex_partners2 ~ sexual_orientation, data = data_self_plus_sexual_partner_variable,
       paired = F)
```

### Plot total number of sex partners and sexual experience againts each other
```{r}
b <- ggplot(data_self_plus_sexual_partner_variable, aes(x = total_sex_partners2, y = self_sexually_experienced))
# Scatter plot with regression line
b + geom_point()+
  geom_smooth(method = "lm") 
```

## Relationship options: look at sexual partners variables
```{r}
# mean
data_rel_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), mean, na.rm = T)

# sd
data_rel_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), sd, na.rm = T)

# min
data_rel_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), min, na.rm = T)

# max
data_rel_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), max, na.rm = T)

# range
data_rel_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), range, na.rm = T)

# median
data_rel_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), median, na.rm = T)

# counts
table(data_rel_plus_sexual_partner_variable$total_sex_partners)
table(data_rel_plus_sexual_partner_variable$total_sex_partners2)
```



## Lifetime sexual partners in unmatched heterosexuals sample
### Load data
```{r}
data_prefs_unmatched = read.csv(file = "data/data_prefs_unmatched.csv")[,-1]
data_self_unmatched = read.csv(file = "data/data_prefs_unmatched.csv")[,-1]

# add sex partner variable to data
data_prefs_unmatched_plus_sexual_partner_variable = left_join(data_prefs_unmatched, sexual_partner)
data_self_unmatched_plus_sexual_partner_variable = left_join(data_self_unmatched, sexual_partner)
```

### Partner preferences
#### Total number of sex partners
```{r}
# mean
data_prefs_unmatched_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), mean, na.rm = T)

# sd
data_prefs_unmatched_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), sd, na.rm = T)

# min
data_prefs_unmatched_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), min, na.rm = T)

# max
data_prefs_unmatched_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), max, na.rm = T)

# range
data_prefs_unmatched_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), range, na.rm = T)

# median
data_prefs_unmatched_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), median, na.rm = T)

# counts
table(data_prefs_unmatched_plus_sexual_partner_variable$total_sex_partners)
table(data_prefs_unmatched_plus_sexual_partner_variable$total_sex_partners2)
```
Using variable total_sex_partners as total_sex_partners2 seems to have unreasonably high values.
Mean number of lifetime sexual partners in unmatched heterosexual sample is 6.23. 

### Self-ratings
#### Total number of sex partners
```{r}
# mean
data_self_unmatched_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), mean, na.rm = T)

# sd
data_self_unmatched_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), sd, na.rm = T)

# min
data_self_unmatched_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), min, na.rm = T)

# max
data_self_unmatched_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), max, na.rm = T)

# range
data_self_unmatched_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), range, na.rm = T)

# median
data_self_unmatched_plus_sexual_partner_variable %>% group_by(sexual_orientation) %>%
  summarise_at(c("total_sex_partners", "total_sex_partners2"), median, na.rm = T)

# counts
table(data_self_unmatched_plus_sexual_partner_variable$total_sex_partners)
table(data_self_unmatched_plus_sexual_partner_variable$total_sex_partners2)
```
Using variable total_sex_partners as total_sex_partners2 seems to have unreasonably high values.
Mean number of lifetime sexual partners in unmatched heterosexual sample is 6.23.
