==> 14_plots.Rmd <==
---
title: "Plots"
output: 
  html_document:
    toc: true
    toc_depth: 6
    toc_float: true
    code_folding: 'show'
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(warning = FALSE)
```

## Library
```{r}
library(dplyr)
library(formr)
library(ggplot2)
library(psych)
library(effsize)
library(forestplot)
library(cobalt)
library(MatchIt)
```

## Load data
```{r}
data_main = read.csv(file = "data/data_main.csv")[,-1]
data_robustness = read.csv(file = "data/data_robustness.csv")[,-1]
```

To create a love plot and a jitter plot, a MatchIt Object is needed. Thus, matching must be rerun.

## Matching for main analyses 
Check for missings in covariates to be matched
```{r}
# MatchIt does not allow missing values
data_main_nomiss <- data_main %>% select(id, language, age, sex, country, relationship_status, relationship_length, sexual_orientation,
                                         pref_imp_ks, pref_imp_att, pref_imp_fs, pref_imp_ca,
                                         pref_imp_ei, pref_imp_sexually_experienced) %>%
                                  na.omit()

table(is.na(data_main_nomiss)) # no more NAs in data set
```

Check if correct variable type 
```{r}
typeof(data_main_nomiss$sexual_orientation) # is variable of type character, but for PSM it needs to be a factor
data_main_nomiss$sexual_orientation = as.factor(data_main_nomiss$sexual_orientation) # convert to factor
levels(data_main_nomiss$sexual_orientation) # check levels of factor
data_main_nomiss$sexual_orientation = factor(data_main_nomiss$sexual_orientation,
                                             levels=rev(levels(data_main_nomiss$sexual_orientation))) # reverse order of factor levels 
levels(data_main_nomiss$sexual_orientation) # check factor level order
```

Matching 
```{r}
nnm <- matchit(sexual_orientation ~ language + age + country + relationship_status + relationship_length, method = "nearest", data = data_main_nomiss)

nnm_data <- match.data(nnm)
```

## Matching for robustness check
```{r}
# MatchIt does not allow missing values
data_robust_nomiss_prefs <- data_robustness %>% select(id, sex, language, age, sex, country, relationship_status, relationship_length, sexual_orientation,
                                         pref_imp_ks, pref_imp_att, pref_imp_fs, pref_imp_ca,
                                         pref_imp_ei, pref_imp_sexually_experienced) %>%
                                  na.omit()

table(is.na(data_robust_nomiss_prefs)) # no more NAs in data set
```

Check if correct variable type 
```{r}
typeof(data_robust_nomiss_prefs$sexual_orientation) # is variable of type character, but for PSM it needs to be a factor
data_robust_nomiss_prefs$sexual_orientation = as.factor(data_robust_nomiss_prefs$sexual_orientation) # convert to factor
levels(data_robust_nomiss_prefs$sexual_orientation) # check levels of factor
data_robust_nomiss_prefs$sexual_orientation = factor(data_robust_nomiss_prefs$sexual_orientation,
                                             levels=rev(levels(data_robust_nomiss_prefs$sexual_orientation))) # reverse order of factor levels 
levels(data_robust_nomiss_prefs$sexual_orientation) # check factor level order
```

Matching with NNM
```{r}
nnm_robust_prefs <- matchit(sexual_orientation ~ sex + language + age + country + relationship_status + relationship_length, method = "nearest", data = data_robust_nomiss_prefs)

nnm_robust_data_prefs <- match.data(nnm_robust_prefs)
```


## 1. Plots for main analyses
### Love plot 
For partner preference sample
```{r}
# omitting country variable for a clearer visualization
nnm_data_without_country = nnm
nnm_data_without_country$X[c("country")] <- NULL 

# Using alternate variable names
v <- data.frame(old = c("distance", "language_chinese", "language_japanese", "language_german","language_english", "language_italian",
                        "language_danish", "language_spanish", "language_russian", "language_portuguese", "language_french", "age",
                        "relationship_status_Long-term", "relationship_status_No relationship", "relationship_status_Short-term",
                        "relationship_length"),
                new = c("Propensity score", "Chinese", "Japanese", "German", "English", "Italian", "Danish", "Spanish", "Russian",
                        "Portuguese", "French", "Long-term relationship", "No relationship", "Dating/Sexual Relationship",
                        "Age",  "Relationship length"))

love.plot(bal.tab(nnm_data_without_country, m.threshold = 0.1, binary = "std"),
          stat = "mean.diffs",
          grid = T,
          stars = "raw",
          abs = F,
          sample.names = c("Before Matching", "After Matching"),
          var.names = v)
```


### Jitter plot
```{r}
plot(nnm, type = "jitter", interactive = FALSE)
```


### Forest plot for each outcome section with effect sizes and 95% CIs
#### Preferred relationship options
```{r, fig.width = 10, fig.height = 6}
base_data_rel <- tibble(mean  = c(-0.67, 1.12, -0.46, 0.20, 0.57, 0.61, -0.77), 
                    lower = c(-0.83, 0.93, -0.61, 0.05, 0.41, 0.45, -0.94),
                    upper = c(-0.51, 1.32, -0.31, 0.35, 0.73, 0.76, -0.61),
                    Preferred_relationship_options = c("Sexual, non-romantic", "Non-sexual, romantic", "Monogamous", "Non-monogamous", "Alternative committed", "Being non-partnered", "Becoming a parent"),
                    d = c("-0.67", "1.12", "-0.46", "0.20", "0.57", "0.61", "-0.77"))


header_rel <- tibble(Preferred_relationship_options = c("", "Outcome"),
                d = c("", "d"),
                 summary = TRUE)

preferred_relationship_options <- bind_rows(header_rel,
                                base_data_rel)



plot_preferred_relationship_options = preferred_relationship_options  %>% 
  forestplot(labeltext = c(Preferred_relationship_options, d), 
             graph.pos = 3,
              clip = c(-1,2),
             lty.ci = 2,
             boxsize = .25,
              is.summary = summary,
              grid = structure(c(-.26, .26), 
                              gp = gpar(lty = 2, col = "darkblue")), 
              xticks = c(-1.5,-1, -0.5, -0.26,0, .26, .5, 1, 1.5),
             xlab = "Effect size (Cohen's d)",
             txt_gp = fpTxtGp(ticks=gpar(cex=.75), xlab=gpar(cex=1)),
             col = fpColors(box = "royalblue",
                            line = "black",
                            summary = "royalblue"),
             title = expression(bold("Preferred Relationship Options ("*bolditalic(N)~"= 646)")),
             graphwidth = unit(150, "mm"),
           colgap = unit(25, "mm")
             )

plot_preferred_relationship_options
```


#### Partner preferences
```{r, fig.width = 10, fig.height = 6}
base_data_prefs <- tibble(mean  = c(-0.37, -0.52, -0.66, -0.16, -0.21, 0.01), 
                    lower = c(-0.51, -0.66, -0.81, -0.30, -0.35, -0.12),
                    upper = c(-0.24, -0.37, -0.51, -0.02, -0.08, 0.15),
                    Partner_preferences = c("Confident-assertive", "Attractive", "Sexually experienced", "Kind-supportive", "Financially secure-successful", "Educated-intelligent"),
                    d = c("-0.37", "-0.52", "-0.66", "-0.16", "-0.21", "0.01"))


header_prefs <- tibble(Partner_preferences = c("", "Outcome"),
                d = c("", "d"),
                 summary = TRUE)


partner_preferences_prefs <- bind_rows(header_prefs,
                                base_data_prefs)



plot_partner_preferences = partner_preferences_prefs %>% 
  forestplot(labeltext = c(Partner_preferences, d), 
             graph.pos = 3,
              clip = c(-2,2),
             lty.ci = 2,
             boxsize = .25,
              is.summary = summary,
              grid = structure(c(-.26, .26), 
                              gp = gpar(lty = 2, col = "darkblue")), 
             xticks = c(-1.5,-1, -0.5, -0.26,0, .26, .5, 1, 1.5),
             xlab = "Effect size (Cohen's d)",
             txt_gp = fpTxtGp(ticks=gpar(cex=.75), xlab = gpar(cex=1)),
             col = fpColors(box = "royalblue",
                            line = "black",
                            summary = "royalblue"),
             title = expression(bold("Ideal Long-Term Partner Preferences ("*bolditalic(N)~"= 780)")),
             graphwidth = unit(150, "mm"))

plot_partner_preferences
```



#### Self-ratings
```{r, fig.width = 10, fig.height = 6}
base_data_self <- tibble(mean  = c(-0.43, -0.49, -0.65, -0.40, -0.45, -0.17), 
                    lower = c(-0.57, -0.63, -0.79, -0.54, -0.60, -0.31),
                    upper = c(-0.29, -0.34, -0.52, -0.25, -0.31, -0.04),
                    self_ratings = c("Confident-assertive", "Attractive", "Sexually experienced", "Kind-supportive", "Financially secure-successful", "Educated-intelligent"),
                    d = c("-0.43", "-0.49", "-0.65", "-0.40", "-0.45", "-0.17"))

header_self <- tibble(self_ratings = c("", "Outcome"),
                d = c("", "d"),
                 summary = TRUE)

self_ratings <- bind_rows(header_self,
                          base_data_self)

plot_self_ratings = self_ratings %>% 
  forestplot(labeltext = c(self_ratings, d), 
             graph.pos = 3,
              clip = c(-2,2),
             lty.ci = 2,
             boxsize = .25,
              is.summary = summary,
              grid = structure(c(-.26, .26), 
                              gp = gpar(lty = 2, col = "darkblue")), 
             xticks = c(-1.5,-1, -0.5, -0.26,0, .26, .5, 1, 1.5),
             xlab = "Effect size (Cohen's d)",
             txt_gp = fpTxtGp(ticks=gpar(cex=.75), xlab=gpar(cex=1)),
             col = fpColors(box = "royalblue",
                            line = "black",
                            summary = "royalblue"),
             title = expression(bold("Self-Ratings ("*bolditalic(N)~"= 772)")),
             graphwidth = unit(150, "mm"))

plot_self_ratings
```

## 2. Plots for robustness check
### Love plot
```{r}
# remove country variable 
nnm_robust_without_country_prefs = nnm_robust_prefs
nnm_robust_without_country_prefs$X[c("country")] <- NULL 

# Using alternate variable names
v <- data.frame(old = c("distance", "language_chinese", "language_japanese", "language_german",
                        "language_english", "language_italian", "language_danish", "language_spanish",
                        "language_russian", "language_portuguese", "language_french",
                        "age", "relationship_status_Long-term", "relationship_status_No relationship",
                        "relationship_status_Short-term", "relationship_length", "sex_Woman", "sex_Man",
                        "sex_Genderqueer/Nonbinary"),
                new = c("Propensity score", "Chinese", "Japanese", "German", "English", "Italian", 
                        "Danish", "Spanish", "Russian", "Portuguese", "French", "Long-term relationship",
                        "No relationship", "Dating/Sexual Relationship","Age",  "Relationship length", 
                        "Woman", "Man", "Genderqueer/Nonbinary"))

love.plot(bal.tab(nnm_robust_without_country_prefs, m.threshold = 0.1, binary = "std"),
          stat = "mean.diffs",
          grid = T,
          stars = "raw",
          abs = F,
          sample.names = c("Before Matching", "After Matching"),
          var.names = v,
          binary = "std") 
```

### Jitter plot
```{r}
plot(nnm_robust_prefs, type = "jitter", interactive = FALSE)
```

### Forest plots for robustness analyses
#### Preferred relationship options
```{r, fig.width = 10, fig.height = 6}
base_data_rel_robust <- tibble(mean  = c(-0.61, 1.12, -0.46, 0.29, 0.61, 0.66, -0.72), 
                               lower = c(-0.76, 0.94, -0.61, 0.15, 0.46, 0.51, -0.87),
                               upper = c(-0.45, 1.29, -0.31, 0.43, 0.76, 0.80, -0.56),
                               Preferred_relationship_options = c("Sexual, non-romantic", "Non-sexual, romantic", "Monogamous", "Non-monogamous", "Alternative committed", "Being non-partnered", "Becoming a parent"),
                               d = c("-0.61", "1.12", "-0.46", "0.29", "0.61", "0.66", "-0.72"))


header_rel_robust <- tibble(Preferred_relationship_options = c("", "Outcome"),
                d = c("", "d"),
                 summary = TRUE)


preferred_relationship_options_robust <- bind_rows(header_rel_robust,
                                base_data_rel_robust)



plot_preferred_relationship_options_robust = preferred_relationship_options_robust  %>% 
  forestplot(labeltext = c(Preferred_relationship_options, d), 
             graph.pos = 3,
              clip = c(-1,2),
             lty.ci = 2,
             boxsize = .25,
              is.summary = summary,
              grid = structure(c(-.26, .26), 
                              gp = gpar(lty = 2, col = "darkblue")), 
              xticks = c(-1.5,-1, -0.5, -0.26,0, .26, .5, 1, 1.5),
             xlab = "Effect size (Cohen's d)",
             txt_gp = fpTxtGp(ticks=gpar(cex=.75), xlab=gpar(cex=1)),
             col = fpColors(box = "royalblue",
                            line = "black",
                            summary = "royalblue"),
             title = expression(bold("Preferred Relationship Options ("*bolditalic(N)~"= 758)")),
             graphwidth = unit(150, "mm"),
             colgap = unit(25, "mm"))

plot_preferred_relationship_options_robust
```


#### Partner preferences
```{r, fig.width = 10, fig.height = 6}
base_data_prefs_robust <- tibble(mean  = c(-0.43, -0.63, -0.70, -0.10, -0.24, -0.05), 
                                 lower = c(-0.55, -0.77, -0.84, -0.23, -0.37, -0.18),
                                 upper = c(-0.30, -0.49, -0.57, -0.03, -0.11, 0.08),
                                 Partner_preferences = c("Confident-assertive", "Attractive", "Sexually experienced", "Kind-supportive", "Financially secure-successful", "Educated-intelligent"),
                                 d = c("-0.43", "-0.63", "-0.70", "-0.10", "-0.24", "-0.05"))


header_prefs_robust <- tibble(Partner_preferences = c("", "Outcome"),
                d = c("", "d"),
                 summary = TRUE)


partner_preferences_prefs_robust <- bind_rows(header_prefs_robust,
                                base_data_prefs_robust)



plot_partner_preferences_robust = partner_preferences_prefs_robust %>% 
  forestplot(labeltext = c(Partner_preferences, d), 
             graph.pos = 3,
              clip = c(-2,2),
             lty.ci = 2,
             boxsize = .25,
              is.summary = summary,
              grid = structure(c(-.26, .26), 
                              gp = gpar(lty = 2, col = "darkblue")), 
             xticks = c(-1.5,-1, -0.5, -0.26,0, .26, .5, 1, 1.5),
             xlab = "Effect size (Cohen's d)",
             txt_gp = fpTxtGp(ticks=gpar(cex=.75), xlab = gpar(cex=1)),
             col = fpColors(box = "royalblue",
                            line = "black",
                            summary = "royalblue"),
             title = expression(bold("Ideal Long-Term Partner Preferences ("*bolditalic(N)~"= 780)")),
             graphwidth = unit(150, "mm"))

plot_partner_preferences_robust
```

#### Self-ratings
```{r, fig.width = 10, fig.height = 6}
base_data_self_robust <- tibble(mean  = c(-0.46, -0.49, -0.65, -0.39, -0.45, -0.18), 
                                lower = c(-0.59, -0.62, -0.78, -0.52, -0.58, -0.31),
                                upper = c(-0.33, -0.35, -0.53, -0.25, -0.32, -0.06),
                                self_ratings = c("Confident-assertive", "Attractive", "Sexually experienced", "Kind-supportive", "Financially secure-successful", "Educated-intelligent"),
                                d = c("-0.46", "-0.49", "-0.65", "-0.39", "-0.45", "-0.18"))

header_self_robust <- tibble(self_ratings = c("", "Outcome"),
                d = c("", "d"),
                 summary = TRUE)

self_ratings_robust <- bind_rows(header_self_robust,
                          base_data_self_robust)

plot_self_ratings_robust = self_ratings_robust %>% 
  forestplot(labeltext = c(self_ratings, d), 
             graph.pos = 3,
              clip = c(-2,2),
             lty.ci = 2,
             boxsize = .25,
              is.summary = summary,
              grid = structure(c(-.26, .26), 
                              gp = gpar(lty = 2, col = "darkblue")), 
             xticks = c(-1.5,-1, -0.5, -0.26,0, .26, .5, 1, 1.5),
             xlab = "Effect size (Cohen's d)",
             txt_gp = fpTxtGp(ticks=gpar(cex=.75), xlab=gpar(cex=1)),
             col = fpColors(box = "royalblue",
                            line = "black",
                            summary = "royalblue"),
             title = expression(bold("Self-Ratings ("*bolditalic(N)~"= 772)")),
             graphwidth = unit(150, "mm"))

plot_self_ratings_robust
```

## 3. Plots for additional analyses
### Forest plots for each additional analysis 
#### 1) Controlling for disinterest in long-term relationship
```{r, fig.width = 10, fig.height = 6}
base_data_add <- tibble(mean  = c(-0.45, -0.49, -0.61, -0.14, -0.50, 0.03), 
                    lower = c(-0.61, -0.66, -0.77, -0.30, -0.66, -0.12),
                    upper = c(-0.30, -0.32, -0.46, 0.02, -0.33, 0.19),
                    Partner_preferences = c("Confident-assertive", "Attractive", "Sexually experienced", "Kind-supportive", "Financially secure-successful", "Educated-intelligent"),
                    d = c("-0.45", "-0.49", "-0.61", "-0.14", "-0.50", "0.03"))

header_add <- tibble(Partner_preferences = c("", "Partner preferences"),
                d = c("", "d"),
                 summary = TRUE)

partner_preferences_add <- bind_rows(header_add,
                                base_data_add)



plot_add1 = partner_preferences_add %>% 
            forestplot(labeltext = c(Partner_preferences, d), 
                       graph.pos = 3,
                        clip = c(-1,2),
                       lty.ci = 2,
                       boxsize = .25,
                        is.summary = summary,
                        grid = structure(c(-.26, .26), 
                                        gp = gpar(lty = 2, col = "darkblue")), 
                       xticks = c(-1.5, -.26, -0.5,0, .26, .5, 1, 1.5),
                       xlab = "Effect size (Cohen's d)",
                       txt_gp = fpTxtGp(ticks=gpar(cex=.75), xlab=gpar(cex=1)),
                       col = fpColors(box = "royalblue",
                                      line = "black",
                                      summary = "royalblue"),
                       title = expression(bold("Controlling for disinterest in long-term relationships ("*bolditalic(N)~"= 636)")),
                       graphwidth = unit(150, "mm"))

plot_add1
```

#### 2) + interest in hook ups
```{r, fig.width = 10, fig.height = 6}
base_data_add2 <- tibble(mean  = c(-0.49, -0.43, -0.37, -0.23, -0.44, -0.04), 
                    lower = c(-0.65, -0.60, -0.52, -0.38, -0.60, -0.19),
                    upper = c(-0.33, -0.27, -0.22, -0.08, -0.28, 0.11),
                    Partner_preferences = c("Confident-assertive", "Attractive", "Sexually experienced", "Kind-supportive", "Financially secure-successful", "Educated-intelligent"),
                    d = c("-0.49", "-0.43", "-0.37", "-0.23", "-0.44", "-0.04"))

header_add2 <- tibble(Partner_preferences = c("", "Partner preferences"),
                d = c("", "d"),
                 summary = TRUE)

partner_preferences_add2 <- bind_rows(header_add2,
                                base_data_add2)



plot_add2 = partner_preferences_add2 %>% 
            forestplot(labeltext = c(Partner_preferences, d), 
                       graph.pos = 3,
                        clip = c(-1,2),
                       lty.ci = 2,
                       boxsize = .25,
                        is.summary = summary,
                        grid = structure(c(-.26, .26), 
                                        gp = gpar(lty = 2, col = "darkblue")), 
                       xticks = c(-1.5, -.26, -0.5,0, .26, .5, 1, 1.5),
                       xlab = "Effect size (Cohen's d)",
                       txt_gp = fpTxtGp(ticks=gpar(cex=.75), xlab=gpar(cex=1)),
                       col = fpColors(box = "royalblue",
                                      line = "black",
                                      summary = "royalblue"),
                       title = expression(bold("+ Interest in purely sexual relationships ("*bolditalic(N)~"= 636)")),
                       graphwidth = unit(150, "mm"))

plot_add2
```

#### 3) + Interest in becoming a parent
```{r, fig.width = 10, fig.height = 6}
base_data_add3 <- tibble(mean  = c(-0.35, -0.43, -0.59, -0.06, -0.33, 0.05), 
                    lower = c(-0.51, -0.59, -0.76, -0.22, -0.49, -0.11),
                    upper = c(-0.20, -0.27, -0.43, 0.10, -0.17, 0.20),
                    Partner_preferences = c("Confident-assertive", "Attractive", "Sexually experienced", "Kind-supportive", "Financially secure-successful", "Educated-intelligent"),
                    d = c("-0.35", "-0.43", "-0.59", "-0.06", "-0.33", "0.05"))

header_add3 <- tibble(Partner_preferences = c("", "Partner preferences"),
                d = c("", "d"),
                 summary = TRUE)

partner_preferences_add3 <- bind_rows(header_add3,
                                base_data_add3)


plot_add3 = partner_preferences_add3 %>% 
            forestplot(labeltext = c(Partner_preferences, d), 
                       graph.pos = 3,
                        clip = c(-1,2),
                       lty.ci = 2,
                       boxsize = .25,
                        is.summary = summary,
                        grid = structure(c(-.26, .26), 
                                        gp = gpar(lty = 2, col = "darkblue")), 
                       xticks = c(-1.5, -.26, -0.5,0, .26, .5, 1, 1.5),
                       xlab = "Effect size (Cohen's d)",
                       txt_gp = fpTxtGp(ticks=gpar(cex=.75), xlab=gpar(cex=1)),
                       col = fpColors(box = "royalblue",
                                      line = "black",
                                      summary = "royalblue"),
                       title = expression(bold("+ Parenting interest ("*bolditalic(N)~" = 634)")),
                       graphwidth = unit(150, "mm"))

plot_add3
```
