==> 15_synthetic_dataset.Rmd <==
---
title: "17_syn_data"
output: html_document
date: "2022-08-31"
---

##Library
```{r}
library(tidyverse)
```

##Load data
```{r}
fulldata_selected = read.csv(file = "data/fulldata_selected.csv")[,-1]
names(fulldata_selected)
```

## Divide data
### Danish data
In the danish dataset the scale for relationship options ranges from 1 to 7 (which is wrong). The danish data is therefore simulated separately
```{r}
fulldata_selected_danish = fulldata_selected %>%
  filter(language == "danish")

fulldata_selected = fulldata_selected %>%
  filter(language != "danish")
```

## Create synthetic dataset
The following procedure is based on https://elifesciences.org/articles/53275 (Quintana, 2020) and https://github.com/dsquintana/synthpop-primer (Quintana, 2020)

### Check wether all required packages are installed
```{r}
ipak <- function(pkg) {
  new.pkg <- pkg[!(pkg %in% installed.packages()[, "Package"])]
  if (length(new.pkg))
    install.packages(new.pkg, dependencies = TRUE)
  sapply(pkg, require, character.only = TRUE)
}

packages <- c("synthpop", "tidyverse", "cowplot", "car",
              "simstudy", "mice", "StatMeasures")
ipak(packages)

#install.packages("https://cran.r-project.org/src/contrib/Archive/StatMeasures/StatMeasures_1.0.tar.gz", repos=NULL, type="source")
```

### Function to simulate correlated data  https://debruine.github.io/faux/index.html
```{r}
rnorm_pre <- function (x, mu = 0, sd = 1, r = 0, empirical = FALSE) 
{
  if (!is.vector(x)) 
    stop("x must be a vector")
  if (!is.numeric(x)) 
    stop("x must be numeric")
  if (length(x) < 3) 
    stop("x must have length > 2")
  n <- length(x)
  if (!empirical) {
    sample_params <- sample_from_pop(n, mu, sd, r)
    mu <- sample_params$mu
    sd <- sample_params$sd
    r <- sample_params$r
  }
  y <- stats::rnorm(n)
  z <- r * scale(x)[, 1] + sqrt(1 - r^2) * scale(stats::resid(stats::lm(y ~ 
                                                                          x)))[, 1]
  yresult <- mu + sd * z
  return(yresult)
}
```

### Select data used for analyses
```{r}
example = fulldata_selected
example_danish = fulldata_selected_danish
```

### Create simulated data
```{r}
example_sim_danish = syn(example_danish, seed = 2019,
                            visit.sequence = c(1, 3:50, 2),
                            maxfaclevels = 181)
example_sim = syn(example, seed = 2019,
                  visit.sequence = c(1, 3:50, 2),
                  maxfaclevels = 181)
```

## Save simulated data
```{r}
example_sim_label <- sdc(example_sim, example, 
                          label = "FAKE_DATA") # Adds a "FAKE_DATA" label
example_sim_label_danish <- sdc(example_sim_danish, example, 
                          label = "FAKE_DATA") # Adds a "FAKE_DATA" label

example_sim_dat <- example_sim_label$syn
example_sim_dat_danish <- example_sim_label_danish$syn

example_sim_dat_full <- rbind(example_sim_dat, example_sim_dat_danish)
save(example_sim_dat_full, file = "data/FAKE_data_asexuality-clue.rdata")
write.csv2(example_sim_dat_full, file = "data/FAKE_data_asexuality-clue.csv")
```
